updating maxwell aircradft
