+++ regression/automatic_regression.py
@@ -40,8 +40,6 @@ modules = [
     'scripts/aerodynamics/aerodynamics.py',
     'scripts/aerodynamics/all_moving_surfaces_vlm.py',
     'scripts/aerodynamics/control_surfaces_vlm.py',
-    'scripts/aerodynamics/sears_haack.py',
-    'scripts/aerodynamics/sideslip_and_rotation_vlm.py',
     'scripts/airfoil_import/airfoil_import_test.py',
     'scripts/airfoil_import/airfoil_interpolation_test.py',
     'scripts/atmosphere/atmosphere.py',
@@ -98,6 +96,8 @@ modules = [
     'scripts/solar_network/solar_network.py',
     'scripts/solar_network/solar_low_fidelity_network.py',
     'scripts/solar_radiation/solar_radiation.py',
+    'scripts/aerodynamics/sears_haack.py',
+    'scripts/aerodynamics/sideslip_and_rotation_vlm.py',
     'scripts/SU2_surrogate/BWB-450.py',
     'scripts/sweeps/test_sweeps.py',
     'scripts/test_input_output/test_xml_read_write.py',
