--- regression/scripts/Vehicles/Airfoils/Clark_y_polar_Re_100000.txt
@@ -1,125 +1,0 @@
-  
-       XFOIL         Version 6.96
-  
- Calculated polar for: CLARK Y AIRFOIL                                 
-  
- 1 1 Reynolds number fixed          Mach number fixed         
-  
- xtrf =   1.000 (top)        1.000 (bottom)  
- Mach =   0.000     Re =     0.100 e 6     Ncrit =   9.000
-  
-   alpha    CL        CD       CDp       CM     Top_Xtr  Bot_Xtr
-  ------ -------- --------- --------- -------- -------- --------
-  -9.000  -0.3474   0.10140   0.09638  -0.0367   1.0000   0.1273
-  -8.750  -0.3809   0.10037   0.09553  -0.0393   1.0000   0.1308
-  -8.500  -0.4214   0.09979   0.09515  -0.0389   1.0000   0.1313
-  -8.250  -0.3699   0.09349   0.08874  -0.0357   1.0000   0.1351
-  -8.000  -0.3685   0.09146   0.08675  -0.0335   1.0000   0.1389
-  -7.750  -0.3862   0.09004   0.08545  -0.0311   1.0000   0.1423
-  -7.500  -0.4177   0.08922   0.08477  -0.0278   1.0000   0.1440
-  -7.250  -0.4616   0.08774   0.08342  -0.0294   1.0000   0.1460
-  -7.000  -0.4818   0.08423   0.07996  -0.0300   1.0000   0.1476
-  -6.750  -0.4620   0.08250   0.07829  -0.0229   1.0000   0.1509
-  -6.500  -0.4660   0.08065   0.07647  -0.0207   1.0000   0.1552
-  -6.250  -0.4964   0.07643   0.07214  -0.0280   1.0000   0.1629
-  -6.000  -0.4869   0.07438   0.07022  -0.0227   1.0000   0.1653
-  -5.750  -0.4833   0.07245   0.06830  -0.0203   1.0000   0.1693
-  -5.500  -0.4804   0.06812   0.06384  -0.0253   0.9986   0.1800
-  -5.250  -0.4492   0.06395   0.05946  -0.0321   0.9922   0.1948
-  -5.000  -0.4217   0.06129   0.05687  -0.0322   0.9869   0.2008
-  -4.750  -0.3727   0.04267   0.03658  -0.0508   0.9819   0.1207
-  -4.500  -0.3344   0.03587   0.02837  -0.0541   0.9768   0.1009
-  -4.250  -0.2952   0.03304   0.02513  -0.0572   0.9721   0.0999
-  -4.000  -0.2584   0.03133   0.02290  -0.0593   0.9662   0.1012
-  -3.750  -0.2209   0.02948   0.02067  -0.0615   0.9603   0.1027
-  -3.500  -0.1775   0.02785   0.01885  -0.0648   0.9562   0.1046
-  -3.250  -0.1467   0.02687   0.01776  -0.0656   0.9484   0.1073
-  -3.000  -0.1049   0.02601   0.01674  -0.0683   0.9429   0.1129
-  -2.750  -0.0711   0.02520   0.01588  -0.0696   0.9359   0.1198
-  -2.500  -0.0332   0.02449   0.01516  -0.0716   0.9294   0.1296
-  -2.250   0.0075   0.02360   0.01442  -0.0740   0.9242   0.1486
-  -2.000   0.0372   0.02287   0.01400  -0.0746   0.9157   0.1943
-  -1.750   0.0798   0.02169   0.01361  -0.0777   0.9111   0.3339
-  -1.500   0.1048   0.02083   0.01363  -0.0774   0.9016   0.5186
-  -1.250   0.1696   0.01958   0.01355  -0.0816   0.8979   1.0000
-  -1.000   0.2100   0.01947   0.01315  -0.0838   0.8869   1.0000
-  -0.750   0.2685   0.01902   0.01241  -0.0888   0.8803   1.0000
-  -0.500   0.2971   0.01899   0.01222  -0.0887   0.8684   1.0000
-  -0.250   0.3365   0.01882   0.01188  -0.0905   0.8612   1.0000
-   0.000   0.3674   0.01875   0.01169  -0.0908   0.8514   1.0000
-   0.250   0.3960   0.01875   0.01157  -0.0906   0.8414   1.0000
-   0.500   0.4341   0.01846   0.01117  -0.0918   0.8343   1.0000
-   0.750   0.4587   0.01851   0.01114  -0.0909   0.8228   1.0000
-   1.000   0.4950   0.01820   0.01074  -0.0917   0.8154   1.0000
-   1.250   0.5218   0.01814   0.01061  -0.0910   0.8041   1.0000
-   1.500   0.5474   0.01813   0.01054  -0.0901   0.7925   1.0000
-   1.750   0.5797   0.01788   0.01022  -0.0901   0.7833   1.0000
-   2.000   0.6085   0.01772   0.01001  -0.0896   0.7720   1.0000
-   2.250   0.6339   0.01769   0.00993  -0.0886   0.7589   1.0000
-   2.500   0.6607   0.01761   0.00981  -0.0877   0.7459   1.0000
-   2.750   0.6885   0.01750   0.00966  -0.0870   0.7330   1.0000
-   3.000   0.7166   0.01739   0.00949  -0.0864   0.7195   1.0000
-   3.250   0.7438   0.01730   0.00936  -0.0855   0.7046   1.0000
-   3.500   0.7699   0.01727   0.00928  -0.0846   0.6881   1.0000
-   3.750   0.7954   0.01729   0.00925  -0.0836   0.6706   1.0000
-   4.000   0.8211   0.01732   0.00924  -0.0826   0.6524   1.0000
-   4.250   0.8471   0.01737   0.00921  -0.0817   0.6336   1.0000
-   4.500   0.8738   0.01746   0.00919  -0.0809   0.6150   1.0000
-   4.750   0.8972   0.01771   0.00940  -0.0798   0.5947   1.0000
-   5.000   0.9213   0.01798   0.00962  -0.0788   0.5753   1.0000
-   5.250   0.9460   0.01829   0.00985  -0.0780   0.5574   1.0000
-   5.500   0.9706   0.01863   0.01012  -0.0772   0.5405   1.0000
-   5.750   0.9949   0.01901   0.01045  -0.0763   0.5244   1.0000
-   6.000   1.0188   0.01941   0.01081  -0.0755   0.5089   1.0000
-   6.250   1.0422   0.01978   0.01114  -0.0745   0.4930   1.0000
-   6.500   1.0650   0.02014   0.01146  -0.0735   0.4768   1.0000
-   6.750   1.0873   0.02051   0.01180  -0.0724   0.4611   1.0000
-   7.000   1.1094   0.02094   0.01224  -0.0713   0.4465   1.0000
-   7.250   1.1315   0.02142   0.01276  -0.0703   0.4331   1.0000
-   7.500   1.1536   0.02191   0.01329  -0.0693   0.4197   1.0000
-   7.750   1.1755   0.02241   0.01382  -0.0683   0.4065   1.0000
-   8.000   1.1971   0.02293   0.01435  -0.0672   0.3929   1.0000
-   8.250   1.2182   0.02347   0.01492  -0.0661   0.3789   1.0000
-   8.500   1.2388   0.02405   0.01551  -0.0649   0.3645   1.0000
-   8.750   1.2581   0.02466   0.01615  -0.0635   0.3492   1.0000
-   9.000   1.2755   0.02531   0.01685  -0.0619   0.3326   1.0000
-   9.250   1.2908   0.02599   0.01760  -0.0599   0.3146   1.0000
-   9.500   1.3041   0.02665   0.01830  -0.0576   0.2954   1.0000
-   9.750   1.3159   0.02731   0.01893  -0.0552   0.2763   1.0000
-  10.000   1.3249   0.02802   0.01965  -0.0524   0.2575   1.0000
-  10.250   1.3309   0.02876   0.02046  -0.0492   0.2393   1.0000
-  10.500   1.3363   0.02954   0.02127  -0.0461   0.2233   1.0000
-  10.750   1.3413   0.03044   0.02219  -0.0430   0.2094   1.0000
-  11.000   1.3473   0.03148   0.02322  -0.0403   0.1972   1.0000
-  11.250   1.3539   0.03261   0.02428  -0.0378   0.1865   1.0000
-  11.500   1.3594   0.03385   0.02564  -0.0355   0.1760   1.0000
-  11.750   1.3649   0.03520   0.02707  -0.0333   0.1663   1.0000
-  12.000   1.3683   0.03666   0.02850  -0.0311   0.1570   1.0000
-  12.250   1.3687   0.03826   0.03022  -0.0290   0.1473   1.0000
-  12.500   1.3694   0.04007   0.03217  -0.0271   0.1381   1.0000
-  12.750   1.3687   0.04207   0.03420  -0.0254   0.1294   1.0000
-  13.000   1.3644   0.04442   0.03666  -0.0238   0.1196   1.0000
-  13.250   1.3568   0.04727   0.03965  -0.0225   0.1086   1.0000
-  13.500   1.3456   0.05070   0.04316  -0.0215   0.0969   1.0000
-  13.750   1.3323   0.05464   0.04712  -0.0208   0.0856   1.0000
-  14.000   1.3196   0.05873   0.05120  -0.0202   0.0761   1.0000
-  14.250   1.3103   0.06255   0.05499  -0.0199   0.0694   1.0000
-  14.500   1.3050   0.06622   0.05879  -0.0194   0.0635   1.0000
-  14.750   1.3011   0.06963   0.06219  -0.0193   0.0596   1.0000
-  15.000   1.2996   0.07289   0.06556  -0.0188   0.0562   1.0000
-  15.250   1.2970   0.07647   0.06929  -0.0189   0.0533   1.0000
-  15.500   1.2952   0.07985   0.07272  -0.0191   0.0508   1.0000
-  15.750   1.2975   0.08261   0.07541  -0.0187   0.0485   1.0000
-  16.000   1.2921   0.08687   0.07995  -0.0194   0.0471   1.0000
-  16.250   1.2873   0.09111   0.08441  -0.0200   0.0458   1.0000
-  16.500   1.2822   0.09546   0.08896  -0.0209   0.0448   1.0000
-  16.750   1.2755   0.10014   0.09382  -0.0221   0.0440   1.0000
-  17.000   1.2673   0.10518   0.09905  -0.0238   0.0434   1.0000
-  17.250   1.2568   0.11077   0.10483  -0.0260   0.0430   1.0000
-  17.500   1.2419   0.11739   0.11167  -0.0291   0.0428   1.0000
-  17.750   1.2181   0.12614   0.12069  -0.0340   0.0430   1.0000
-  18.000   1.1716   0.14084   0.13576  -0.0436   0.0444   1.0000
-  18.250   1.0657   0.17534   0.17052  -0.0663   0.0498   1.0000
-  18.500   1.0634   0.18080   0.17597  -0.0691   0.0489   1.0000
-  18.750   0.9950   0.21955   0.21443  -0.0905   0.0654   1.0000
-  19.000   1.0031   0.22395   0.21884  -0.0919   0.0673   1.0000
