--- regression/scripts/Vehicles/Airfoils/Clark_y_polar_Re_200000.txt
@@ -1,114 +1,0 @@
-  
-       XFOIL         Version 6.96
-  
- Calculated polar for: CLARK Y AIRFOIL                                 
-  
- 1 1 Reynolds number fixed          Mach number fixed         
-  
- xtrf =   1.000 (top)        1.000 (bottom)  
- Mach =   0.000     Re =     0.200 e 6     Ncrit =   9.000
-  
-   alpha    CL        CD       CDp       CM     Top_Xtr  Bot_Xtr
-  ------ -------- --------- --------- -------- -------- --------
-  -9.000  -0.4082   0.09344   0.09005  -0.0468   1.0000   0.0739
-  -8.750  -0.4391   0.09151   0.08824  -0.0450   1.0000   0.0740
-  -8.500  -0.4670   0.08892   0.08571  -0.0443   1.0000   0.0741
-  -8.250  -0.4772   0.08456   0.08142  -0.0411   1.0000   0.0748
-  -8.000  -0.4670   0.08349   0.08039  -0.0352   1.0000   0.0758
-  -7.750  -0.4730   0.08209   0.07903  -0.0319   1.0000   0.0765
-  -7.500  -0.4528   0.07915   0.07607  -0.0338   0.9969   0.0785
-  -7.250  -0.4291   0.07367   0.07053  -0.0427   0.9910   0.0830
-  -7.000  -0.4158   0.06100   0.05751  -0.0619   0.9808   0.0895
-  -6.750  -0.3903   0.05905   0.05559  -0.0625   0.9760   0.0913
-  -6.500  -0.3615   0.05546   0.05190  -0.0668   0.9718   0.0951
-  -6.250  -0.3450   0.04808   0.04400  -0.0741   0.9623   0.1041
-  -6.000  -0.3240   0.03416   0.02886  -0.0789   0.9571   0.0704
-  -5.750  -0.3036   0.02800   0.02149  -0.0788   0.9496   0.0631
-  -5.500  -0.2702   0.02569   0.01892  -0.0806   0.9455   0.0638
-  -5.250  -0.2318   0.02391   0.01690  -0.0831   0.9428   0.0648
-  -5.000  -0.1910   0.02223   0.01497  -0.0858   0.9410   0.0653
-  -4.750  -0.1654   0.02118   0.01377  -0.0854   0.9327   0.0660
-  -4.500  -0.1273   0.02005   0.01249  -0.0874   0.9294   0.0673
-  -4.250  -0.0881   0.01908   0.01142  -0.0895   0.9268   0.0695
-  -4.000  -0.0604   0.01847   0.01072  -0.0893   0.9193   0.0719
-  -3.750  -0.0261   0.01778   0.00991  -0.0903   0.9145   0.0739
-  -3.500   0.0091   0.01659   0.00877  -0.0916   0.9112   0.0772
-  -3.250   0.0342   0.01609   0.00829  -0.0910   0.9027   0.0811
-  -3.000   0.0668   0.01553   0.00768  -0.0915   0.8975   0.0871
-  -2.750   0.0940   0.01485   0.00710  -0.0912   0.8900   0.0985
-  -2.500   0.1221   0.01403   0.00642  -0.0910   0.8825   0.1304
-  -2.250   0.1479   0.01328   0.00605  -0.0905   0.8741   0.2081
-  -2.000   0.1751   0.01271   0.00578  -0.0901   0.8662   0.2941
-  -1.750   0.1998   0.01226   0.00560  -0.0894   0.8568   0.3729
-  -1.500   0.2249   0.01156   0.00536  -0.0885   0.8491   0.4994
-  -1.250   0.2429   0.01087   0.00534  -0.0860   0.8383   0.6774
-  -1.000   0.2709   0.01036   0.00534  -0.0841   0.8309   0.8768
-  -0.750   0.3315   0.01030   0.00523  -0.0899   0.8231   0.9730
-  -0.500   0.3953   0.01014   0.00489  -0.0970   0.8170   1.0000
-  -0.250   0.4182   0.01015   0.00480  -0.0960   0.8055   1.0000
-   0.000   0.4424   0.01016   0.00470  -0.0952   0.7954   1.0000
-   0.250   0.4680   0.01015   0.00455  -0.0945   0.7864   1.0000
-   0.500   0.4915   0.01020   0.00453  -0.0935   0.7750   1.0000
-   0.750   0.5161   0.01024   0.00447  -0.0927   0.7644   1.0000
-   1.000   0.5419   0.01026   0.00437  -0.0920   0.7544   1.0000
-   1.250   0.5656   0.01032   0.00437  -0.0911   0.7423   1.0000
-   1.500   0.5899   0.01039   0.00437  -0.0902   0.7308   1.0000
-   1.750   0.6149   0.01046   0.00436  -0.0895   0.7196   1.0000
-   2.000   0.6402   0.01053   0.00433  -0.0887   0.7080   1.0000
-   2.250   0.6641   0.01061   0.00436  -0.0877   0.6945   1.0000
-   2.500   0.6883   0.01071   0.00440  -0.0868   0.6807   1.0000
-   2.750   0.7126   0.01081   0.00444  -0.0859   0.6666   1.0000
-   3.000   0.7370   0.01092   0.00449  -0.0851   0.6519   1.0000
-   3.250   0.7612   0.01104   0.00456  -0.0842   0.6360   1.0000
-   3.500   0.7852   0.01117   0.00462  -0.0832   0.6187   1.0000
-   3.750   0.8091   0.01133   0.00468  -0.0823   0.6003   1.0000
-   4.000   0.8324   0.01152   0.00480  -0.0812   0.5801   1.0000
-   4.250   0.8553   0.01174   0.00494  -0.0801   0.5580   1.0000
-   4.500   0.8781   0.01201   0.00508  -0.0790   0.5358   1.0000
-   4.750   0.9005   0.01231   0.00530  -0.0779   0.5124   1.0000
-   5.000   0.9227   0.01266   0.00551  -0.0768   0.4911   1.0000
-   5.250   0.9450   0.01300   0.00576  -0.0758   0.4700   1.0000
-   5.500   0.9671   0.01335   0.00602  -0.0747   0.4502   1.0000
-   5.750   0.9892   0.01371   0.00630  -0.0737   0.4321   1.0000
-   6.000   1.0116   0.01408   0.00660  -0.0727   0.4164   1.0000
-   6.250   1.0345   0.01444   0.00693  -0.0719   0.4033   1.0000
-   6.500   1.0576   0.01483   0.00730  -0.0711   0.3917   1.0000
-   6.750   1.0804   0.01525   0.00766  -0.0703   0.3808   1.0000
-   7.000   1.1032   0.01563   0.00805  -0.0695   0.3699   1.0000
-   7.250   1.1262   0.01602   0.00848  -0.0687   0.3598   1.0000
-   7.500   1.1486   0.01648   0.00890  -0.0679   0.3499   1.0000
-   7.750   1.1705   0.01686   0.00933  -0.0669   0.3391   1.0000
-   8.000   1.1918   0.01726   0.00978  -0.0659   0.3278   1.0000
-   8.250   1.2120   0.01769   0.01023  -0.0647   0.3155   1.0000
-   8.500   1.2310   0.01811   0.01066  -0.0634   0.3022   1.0000
-   8.750   1.2490   0.01852   0.01110  -0.0618   0.2876   1.0000
-   9.000   1.2663   0.01891   0.01155  -0.0602   0.2720   1.0000
-   9.250   1.2826   0.01932   0.01203  -0.0584   0.2533   1.0000
-   9.500   1.2960   0.01984   0.01253  -0.0563   0.2317   1.0000
-   9.750   1.3072   0.02048   0.01314  -0.0538   0.2074   1.0000
-  10.000   1.3158   0.02138   0.01392  -0.0512   0.1848   1.0000
-  10.250   1.3246   0.02239   0.01486  -0.0487   0.1663   1.0000
-  10.500   1.3333   0.02347   0.01590  -0.0464   0.1524   1.0000
-  10.750   1.3420   0.02458   0.01699  -0.0442   0.1408   1.0000
-  11.000   1.3494   0.02581   0.01818  -0.0421   0.1314   1.0000
-  11.250   1.3609   0.02682   0.01927  -0.0404   0.1222   1.0000
-  11.500   1.3704   0.02799   0.02047  -0.0387   0.1126   1.0000
-  11.750   1.3775   0.02935   0.02184  -0.0370   0.1040   1.0000
-  12.000   1.3859   0.03069   0.02323  -0.0355   0.0937   1.0000
-  12.250   1.3920   0.03227   0.02483  -0.0339   0.0811   1.0000
-  12.500   1.3965   0.03407   0.02662  -0.0324   0.0629   1.0000
-  12.750   1.3940   0.03655   0.02899  -0.0307   0.0468   1.0000
-  13.000   1.3907   0.03921   0.03163  -0.0292   0.0389   1.0000
-  13.250   1.3873   0.04198   0.03444  -0.0279   0.0350   1.0000
-  13.500   1.3833   0.04495   0.03749  -0.0269   0.0328   1.0000
-  13.750   1.3802   0.04793   0.04060  -0.0262   0.0311   1.0000
-  14.000   1.3753   0.05122   0.04401  -0.0257   0.0299   1.0000
-  14.250   1.3681   0.05491   0.04780  -0.0255   0.0289   1.0000
-  14.500   1.3581   0.05908   0.05208  -0.0256   0.0281   1.0000
-  14.750   1.3508   0.06309   0.05620  -0.0259   0.0275   1.0000
-  15.000   1.3451   0.06702   0.06027  -0.0263   0.0268   1.0000
-  15.250   1.3390   0.07109   0.06447  -0.0269   0.0262   1.0000
-  15.500   1.3327   0.07526   0.06876  -0.0276   0.0255   1.0000
-  15.750   1.3262   0.07955   0.07316  -0.0285   0.0250   1.0000
-  16.000   1.3199   0.08387   0.07756  -0.0294   0.0244   1.0000
-  16.250   1.3136   0.08815   0.08192  -0.0303   0.0239   1.0000
