--- regression/scripts/Vehicles/Airfoils/Clark_y_polar_Re_50000.txt
@@ -1,124 +1,0 @@
-  
-       XFOIL         Version 6.96
-  
- Calculated polar for: CLARK Y AIRFOIL                                 
-  
- 1 1 Reynolds number fixed          Mach number fixed         
-  
- xtrf =   1.000 (top)        1.000 (bottom)  
- Mach =   0.000     Re =     0.050 e 6     Ncrit =   5.000
-  
-   alpha    CL        CD       CDp       CM     Top_Xtr  Bot_Xtr
-  ------ -------- --------- --------- -------- -------- --------
- -10.750  -0.3683   0.12749   0.11999  -0.0321   1.0000   0.1294
- -10.500  -0.3799   0.12552   0.11811  -0.0349   1.0000   0.1311
- -10.250  -0.3871   0.12285   0.11552  -0.0371   1.0000   0.1315
- -10.000  -0.3561   0.11742   0.11002  -0.0338   1.0000   0.1362
-  -9.750  -0.3865   0.10888   0.10159  -0.0427   1.0000   0.0878
-  -9.500  -0.3694   0.10555   0.09820  -0.0406   1.0000   0.0851
-  -9.250  -0.3674   0.10205   0.09476  -0.0408   1.0000   0.0829
-  -9.000  -0.3705   0.09846   0.09122  -0.0415   1.0000   0.0808
-  -8.750  -0.3776   0.09491   0.08777  -0.0421   1.0000   0.0795
-  -8.250  -0.4394   0.08464   0.07783  -0.0474   1.0000   0.0740
-  -8.000  -0.4506   0.08108   0.07432  -0.0472   1.0000   0.0739
-  -7.750  -0.4626   0.07740   0.07067  -0.0469   1.0000   0.0738
-  -7.500  -0.4743   0.07361   0.06686  -0.0465   1.0000   0.0738
-  -7.250  -0.4843   0.06974   0.06291  -0.0461   1.0000   0.0738
-  -7.000  -0.4913   0.06583   0.05888  -0.0457   1.0000   0.0738
-  -6.750  -0.4944   0.06200   0.05487  -0.0453   1.0000   0.0737
-  -6.500  -0.4936   0.05827   0.05092  -0.0448   1.0000   0.0736
-  -6.250  -0.4896   0.05457   0.04694  -0.0445   1.0000   0.0735
-  -6.000  -0.4821   0.05100   0.04301  -0.0443   1.0000   0.0736
-  -5.750  -0.4540   0.04675   0.03823  -0.0478   0.9942   0.0746
-  -5.500  -0.4248   0.04466   0.03605  -0.0499   0.9882   0.0770
-  -5.250  -0.3926   0.04209   0.03312  -0.0527   0.9827   0.0793
-  -5.000  -0.3627   0.03934   0.02988  -0.0547   0.9761   0.0807
-  -4.750  -0.3280   0.03687   0.02688  -0.0571   0.9709   0.0823
-  -4.500  -0.2966   0.03491   0.02444  -0.0586   0.9642   0.0845
-  -4.250  -0.2616   0.03329   0.02236  -0.0605   0.9584   0.0884
-  -4.000  -0.2293   0.03213   0.02111  -0.0619   0.9522   0.0922
-  -3.750  -0.1962   0.03099   0.01975  -0.0632   0.9456   0.0959
-  -3.500  -0.1577   0.02990   0.01834  -0.0652   0.9408   0.1008
-  -3.250  -0.1300   0.02912   0.01758  -0.0655   0.9325   0.1071
-  -3.000  -0.0927   0.02836   0.01668  -0.0674   0.9270   0.1181
-  -2.750  -0.0630   0.02772   0.01598  -0.0679   0.9193   0.1302
-  -2.500  -0.0276   0.02702   0.01537  -0.0696   0.9130   0.1521
-  -2.250   0.0039   0.02644   0.01494  -0.0707   0.9057   0.1918
-  -2.000   0.0363   0.02580   0.01472  -0.0721   0.8985   0.2686
-  -1.750   0.0685   0.02503   0.01447  -0.0733   0.8921   0.3781
-  -1.500   0.0939   0.02427   0.01439  -0.0728   0.8840   0.5183
-  -1.250   0.1299   0.02332   0.01447  -0.0719   0.8799   0.8331
-  -1.000   0.1803   0.02334   0.01424  -0.0763   0.8714   1.0000
-  -0.750   0.2176   0.02339   0.01398  -0.0781   0.8645   1.0000
-  -0.500   0.2416   0.02361   0.01395  -0.0777   0.8536   1.0000
-  -0.250   0.2742   0.02368   0.01379  -0.0785   0.8446   1.0000
-   0.000   0.3084   0.02361   0.01351  -0.0793   0.8341   1.0000
-   0.250   0.3367   0.02360   0.01332  -0.0790   0.8210   1.0000
-   0.500   0.3660   0.02355   0.01311  -0.0789   0.8082   1.0000
-   0.750   0.3972   0.02346   0.01288  -0.0789   0.7967   1.0000
-   1.000   0.4301   0.02334   0.01263  -0.0793   0.7868   1.0000
-   1.250   0.4543   0.02348   0.01268  -0.0785   0.7743   1.0000
-   1.500   0.4808   0.02357   0.01268  -0.0779   0.7626   1.0000
-   1.750   0.5119   0.02350   0.01252  -0.0780   0.7527   1.0000
-   2.000   0.5391   0.02355   0.01250  -0.0775   0.7408   1.0000
-   2.250   0.5637   0.02369   0.01260  -0.0766   0.7279   1.0000
-   2.500   0.5897   0.02377   0.01264  -0.0759   0.7152   1.0000
-   2.750   0.6171   0.02381   0.01263  -0.0754   0.7027   1.0000
-   3.000   0.6462   0.02377   0.01256  -0.0750   0.6908   1.0000
-   3.250   0.6728   0.02383   0.01260  -0.0743   0.6773   1.0000
-   3.500   0.6976   0.02396   0.01272  -0.0734   0.6626   1.0000
-   3.750   0.7225   0.02410   0.01286  -0.0725   0.6478   1.0000
-   4.000   0.7476   0.02425   0.01300  -0.0717   0.6327   1.0000
-   4.250   0.7728   0.02441   0.01315  -0.0708   0.6173   1.0000
-   4.500   0.7979   0.02459   0.01334  -0.0700   0.6017   1.0000
-   4.750   0.8229   0.02481   0.01355  -0.0691   0.5861   1.0000
-   5.000   0.8478   0.02506   0.01380  -0.0683   0.5704   1.0000
-   5.250   0.8724   0.02535   0.01411  -0.0675   0.5549   1.0000
-   5.500   0.8967   0.02571   0.01447  -0.0667   0.5398   1.0000
-   5.750   0.9206   0.02611   0.01489  -0.0658   0.5250   1.0000
-   6.000   0.9442   0.02655   0.01537  -0.0650   0.5105   1.0000
-   6.250   0.9676   0.02703   0.01589  -0.0642   0.4965   1.0000
-   6.500   0.9908   0.02753   0.01642  -0.0633   0.4828   1.0000
-   6.750   1.0137   0.02803   0.01695  -0.0624   0.4694   1.0000
-   7.000   1.0365   0.02854   0.01751  -0.0615   0.4560   1.0000
-   7.250   1.0589   0.02904   0.01803  -0.0605   0.4426   1.0000
-   7.500   1.0802   0.02956   0.01859  -0.0593   0.4286   1.0000
-   7.750   1.0990   0.03014   0.01923  -0.0579   0.4135   1.0000
-   8.000   1.1171   0.03072   0.01985  -0.0563   0.3981   1.0000
-   8.250   1.1344   0.03136   0.02052  -0.0547   0.3829   1.0000
-   8.500   1.1507   0.03207   0.02129  -0.0530   0.3680   1.0000
-   8.750   1.1663   0.03284   0.02218  -0.0513   0.3538   1.0000
-   9.000   1.1812   0.03368   0.02312  -0.0496   0.3398   1.0000
-   9.250   1.1953   0.03458   0.02412  -0.0478   0.3261   1.0000
-   9.500   1.2082   0.03551   0.02516  -0.0459   0.3122   1.0000
-   9.750   1.2198   0.03650   0.02626  -0.0438   0.2985   1.0000
-  10.000   1.2292   0.03753   0.02741  -0.0415   0.2851   1.0000
-  10.250   1.2370   0.03863   0.02860  -0.0392   0.2718   1.0000
-  10.500   1.2441   0.03980   0.02984  -0.0368   0.2588   1.0000
-  10.750   1.2481   0.04125   0.03147  -0.0345   0.2459   1.0000
-  11.000   1.2524   0.04284   0.03320  -0.0324   0.2338   1.0000
-  11.250   1.2570   0.04447   0.03496  -0.0304   0.2226   1.0000
-  11.500   1.2624   0.04606   0.03662  -0.0286   0.2125   1.0000
-  11.750   1.2666   0.04784   0.03850  -0.0269   0.2028   1.0000
-  12.000   1.2690   0.04994   0.04076  -0.0254   0.1939   1.0000
-  12.250   1.2747   0.05164   0.04246  -0.0240   0.1859   1.0000
-  12.500   1.2723   0.05423   0.04524  -0.0227   0.1775   1.0000
-  12.750   1.2707   0.05664   0.04770  -0.0216   0.1694   1.0000
-  13.000   1.2669   0.05930   0.05044  -0.0208   0.1611   1.0000
-  13.250   1.2598   0.06261   0.05389  -0.0203   0.1535   1.0000
-  13.500   1.2563   0.06533   0.05660  -0.0198   0.1456   1.0000
-  13.750   1.2448   0.06962   0.06115  -0.0201   0.1389   1.0000
-  14.000   1.2399   0.07271   0.06416  -0.0202   0.1313   1.0000
-  14.250   1.2245   0.07794   0.06970  -0.0214   0.1250   1.0000
-  14.500   1.2176   0.08161   0.07330  -0.0221   0.1177   1.0000
-  14.750   1.2014   0.08752   0.07952  -0.0240   0.1120   1.0000
-  15.000   1.1960   0.09127   0.08321  -0.0250   0.1052   1.0000
-  15.250   1.1795   0.09779   0.09003  -0.0275   0.1005   1.0000
-  15.500   1.1711   0.10264   0.09495  -0.0292   0.0944   1.0000
-  15.750   1.1586   0.10858   0.10103  -0.0316   0.0893   1.0000
-  16.000   1.1435   0.11533   0.10796  -0.0346   0.0844   1.0000
-  16.250   1.1380   0.11991   0.11252  -0.0365   0.0785   1.0000
-  16.500   1.1148   0.12918   0.12205  -0.0412   0.0752   1.0000
-  16.750   1.1211   0.13104   0.12374  -0.0416   0.0685   1.0000
-  17.000   1.0920   0.14257   0.13553  -0.0479   0.0668   1.0000
-  17.250   1.0525   0.15811   0.15118  -0.0568   0.0657   1.0000
