--- regression/scripts/Vehicles/Airfoils/Clark_y_polar_Re_500000.txt
@@ -1,118 +1,0 @@
-  
-       XFOIL         Version 6.96
-  
- Calculated polar for: CLARK Y AIRFOIL                                 
-  
- 1 1 Reynolds number fixed          Mach number fixed         
-  
- xtrf =   1.000 (top)        1.000 (bottom)  
- Mach =   0.000     Re =     0.500 e 6     Ncrit =   9.000
-  
-   alpha    CL        CD       CDp       CM     Top_Xtr  Bot_Xtr
-  ------ -------- --------- --------- -------- -------- --------
-  -9.250  -0.3940   0.08504   0.08280  -0.0444   1.0000   0.0381
-  -9.000  -0.4033   0.08275   0.08056  -0.0431   1.0000   0.0387
-  -8.750  -0.4929   0.04594   0.04332  -0.0802   0.9855   0.0348
-  -8.500  -0.4974   0.03456   0.03116  -0.0866   0.9772   0.0344
-  -8.250  -0.4759   0.02994   0.02595  -0.0893   0.9733   0.0352
-  -8.000  -0.4463   0.02768   0.02325  -0.0914   0.9710   0.0358
-  -7.750  -0.4301   0.02331   0.01843  -0.0916   0.9643   0.0365
-  -7.500  -0.4024   0.02155   0.01653  -0.0925   0.9600   0.0373
-  -7.250  -0.3722   0.02018   0.01500  -0.0935   0.9568   0.0379
-  -7.000  -0.3479   0.01919   0.01389  -0.0931   0.9498   0.0386
-  -6.750  -0.3205   0.01829   0.01288  -0.0932   0.9443   0.0394
-  -6.500  -0.2942   0.01724   0.01167  -0.0930   0.9387   0.0399
-  -6.250  -0.2696   0.01629   0.01057  -0.0923   0.9311   0.0403
-  -6.000  -0.2426   0.01543   0.00958  -0.0921   0.9257   0.0408
-  -5.750  -0.2181   0.01471   0.00876  -0.0913   0.9173   0.0413
-  -5.500  -0.1915   0.01408   0.00802  -0.0910   0.9110   0.0419
-  -5.250  -0.1658   0.01361   0.00747  -0.0904   0.9029   0.0426
-  -5.000  -0.1391   0.01312   0.00689  -0.0900   0.8960   0.0430
-  -4.750  -0.1139   0.01244   0.00613  -0.0894   0.8877   0.0436
-  -4.500  -0.0885   0.01171   0.00534  -0.0889   0.8804   0.0446
-  -4.250  -0.0629   0.01125   0.00485  -0.0884   0.8713   0.0457
-  -4.000  -0.0363   0.01091   0.00444  -0.0879   0.8630   0.0469
-  -3.750  -0.0102   0.01061   0.00411  -0.0875   0.8528   0.0482
-  -3.500   0.0167   0.01036   0.00381  -0.0871   0.8441   0.0499
-  -3.250   0.0438   0.01016   0.00354  -0.0868   0.8350   0.0516
-  -3.000   0.0703   0.00985   0.00319  -0.0864   0.8255   0.0547
-  -2.750   0.0975   0.00965   0.00296  -0.0861   0.8168   0.0589
-  -2.500   0.1243   0.00942   0.00275  -0.0857   0.8067   0.0669
-  -2.250   0.1510   0.00915   0.00254  -0.0854   0.7976   0.0936
-  -2.000   0.1774   0.00887   0.00238  -0.0851   0.7876   0.1350
-  -1.750   0.2037   0.00859   0.00228  -0.0848   0.7769   0.1935
-  -1.500   0.2303   0.00838   0.00221  -0.0846   0.7675   0.2491
-  -1.250   0.2571   0.00820   0.00215  -0.0844   0.7583   0.2986
-  -1.000   0.2836   0.00799   0.00210  -0.0841   0.7496   0.3576
-  -0.750   0.3095   0.00774   0.00206  -0.0838   0.7404   0.4367
-  -0.500   0.3338   0.00732   0.00205  -0.0831   0.7305   0.5667
-  -0.250   0.3551   0.00684   0.00209  -0.0816   0.7213   0.7347
-   0.000   0.3760   0.00652   0.00219  -0.0794   0.7115   0.8796
-   0.250   0.4107   0.00654   0.00224  -0.0804   0.7020   0.9451
-   0.500   0.4534   0.00663   0.00225  -0.0833   0.6921   0.9765
-   0.750   0.5065   0.00669   0.00224  -0.0886   0.6803   0.9953
-   1.000   0.5439   0.00675   0.00222  -0.0907   0.6684   1.0000
-   1.250   0.5687   0.00682   0.00222  -0.0901   0.6567   1.0000
-   1.500   0.5934   0.00691   0.00222  -0.0894   0.6450   1.0000
-   1.750   0.6180   0.00699   0.00225  -0.0887   0.6323   1.0000
-   2.000   0.6426   0.00708   0.00228  -0.0880   0.6193   1.0000
-   2.250   0.6670   0.00719   0.00232  -0.0872   0.6053   1.0000
-   2.500   0.6912   0.00732   0.00237  -0.0864   0.5901   1.0000
-   2.750   0.7151   0.00746   0.00244  -0.0855   0.5739   1.0000
-   3.000   0.7389   0.00761   0.00251  -0.0847   0.5572   1.0000
-   3.250   0.7624   0.00779   0.00261  -0.0838   0.5390   1.0000
-   3.500   0.7857   0.00799   0.00271  -0.0828   0.5188   1.0000
-   3.750   0.8090   0.00820   0.00284  -0.0819   0.4959   1.0000
-   4.000   0.8315   0.00848   0.00299  -0.0808   0.4690   1.0000
-   4.250   0.8536   0.00880   0.00315  -0.0797   0.4386   1.0000
-   4.500   0.8757   0.00914   0.00334  -0.0787   0.4111   1.0000
-   4.750   0.8987   0.00945   0.00355  -0.0778   0.3891   1.0000
-   5.000   0.9220   0.00976   0.00376  -0.0770   0.3723   1.0000
-   5.250   0.9457   0.01004   0.00398  -0.0763   0.3590   1.0000
-   5.500   0.9697   0.01032   0.00421  -0.0756   0.3484   1.0000
-   5.750   0.9940   0.01057   0.00444  -0.0750   0.3396   1.0000
-   6.000   1.0185   0.01082   0.00468  -0.0745   0.3319   1.0000
-   6.250   1.0426   0.01108   0.00492  -0.0739   0.3237   1.0000
-   6.500   1.0671   0.01132   0.00517  -0.0733   0.3159   1.0000
-   6.750   1.0910   0.01160   0.00543  -0.0727   0.3073   1.0000
-   7.000   1.1155   0.01183   0.00567  -0.0721   0.2979   1.0000
-   7.250   1.1388   0.01212   0.00594  -0.0715   0.2876   1.0000
-   7.500   1.1627   0.01238   0.00619  -0.0708   0.2767   1.0000
-   7.750   1.1866   0.01263   0.00644  -0.0702   0.2648   1.0000
-   8.000   1.2096   0.01293   0.00672  -0.0695   0.2509   1.0000
-   8.250   1.2317   0.01328   0.00703  -0.0687   0.2337   1.0000
-   8.500   1.2520   0.01375   0.00740  -0.0676   0.2092   1.0000
-   8.750   1.2691   0.01442   0.00789  -0.0660   0.1778   1.0000
-   9.000   1.2847   0.01519   0.00850  -0.0643   0.1507   1.0000
-   9.250   1.2999   0.01594   0.00914  -0.0624   0.1304   1.0000
-   9.500   1.3150   0.01659   0.00973  -0.0605   0.1159   1.0000
-   9.750   1.3299   0.01726   0.01035  -0.0587   0.1043   1.0000
-  10.000   1.3447   0.01794   0.01099  -0.0568   0.0940   1.0000
-  10.250   1.3602   0.01858   0.01163  -0.0551   0.0853   1.0000
-  10.500   1.3747   0.01929   0.01232  -0.0534   0.0759   1.0000
-  10.750   1.3868   0.02016   0.01312  -0.0514   0.0621   1.0000
-  11.000   1.3917   0.02153   0.01429  -0.0487   0.0373   1.0000
-  11.250   1.3923   0.02324   0.01587  -0.0456   0.0223   1.0000
-  11.500   1.3998   0.02452   0.01720  -0.0434   0.0198   1.0000
-  11.750   1.4079   0.02581   0.01857  -0.0415   0.0183   1.0000
-  12.000   1.4163   0.02711   0.01997  -0.0397   0.0173   1.0000
-  12.250   1.4223   0.02865   0.02159  -0.0380   0.0164   1.0000
-  12.500   1.4251   0.03050   0.02353  -0.0361   0.0157   1.0000
-  12.750   1.4237   0.03282   0.02595  -0.0343   0.0151   1.0000
-  13.000   1.4287   0.03464   0.02789  -0.0330   0.0148   1.0000
-  13.250   1.4317   0.03673   0.03008  -0.0319   0.0145   1.0000
-  13.500   1.4329   0.03907   0.03254  -0.0308   0.0141   1.0000
-  13.750   1.4323   0.04167   0.03524  -0.0299   0.0138   1.0000
-  14.000   1.4303   0.04452   0.03820  -0.0293   0.0135   1.0000
-  14.250   1.4269   0.04763   0.04141  -0.0288   0.0133   1.0000
-  14.500   1.4220   0.05103   0.04492  -0.0285   0.0130   1.0000
-  14.750   1.4153   0.05481   0.04879  -0.0286   0.0128   1.0000
-  15.000   1.4066   0.05897   0.05306  -0.0288   0.0126   1.0000
-  15.250   1.3959   0.06349   0.05768  -0.0293   0.0124   1.0000
-  15.500   1.3832   0.06838   0.06268  -0.0300   0.0122   1.0000
-  15.750   1.3730   0.07302   0.06742  -0.0307   0.0120   1.0000
-  16.000   1.3700   0.07690   0.07141  -0.0315   0.0119   1.0000
-  16.250   1.3662   0.08095   0.07557  -0.0325   0.0117   1.0000
-  16.500   1.3615   0.08517   0.07990  -0.0335   0.0116   1.0000
-  16.750   1.3564   0.08948   0.08431  -0.0346   0.0114   1.0000
-  17.000   1.3510   0.09382   0.08876  -0.0358   0.0113   1.0000
