--- regression/scripts/Vehicles/Airfoils/NACA_4412_polar_Re_100000.txt
@@ -1,119 +1,0 @@
-  
-       XFOIL         Version 6.96
-  
- Calculated polar for: NACA 4412                                       
-  
- 1 1 Reynolds number fixed          Mach number fixed         
-  
- xtrf =   1.000 (top)        1.000 (bottom)  
- Mach =   0.000     Re =     0.100 e 6     Ncrit =   9.000
-  
-   alpha    CL        CD       CDp       CM     Top_Xtr  Bot_Xtr
-  ------ -------- --------- --------- -------- -------- --------
-  -9.500  -0.3426   0.10705   0.10189  -0.0379   1.0000   0.1297
-  -9.250  -0.3784   0.10671   0.10171  -0.0397   1.0000   0.1327
-  -9.000  -0.4173   0.10641   0.10160  -0.0391   1.0000   0.1332
-  -8.750  -0.3682   0.09949   0.09457  -0.0361   1.0000   0.1363
-  -8.500  -0.3611   0.09726   0.09235  -0.0338   1.0000   0.1402
-  -8.250  -0.3724   0.09561   0.09079  -0.0319   1.0000   0.1442
-  -8.000  -0.4032   0.09481   0.09013  -0.0296   1.0000   0.1469
-  -7.750  -0.4436   0.09403   0.08951  -0.0287   1.0000   0.1481
-  -7.500  -0.4819   0.09082   0.08639  -0.0339   1.0000   0.1493
-  -7.250  -0.4471   0.08830   0.08388  -0.0235   1.0000   0.1526
-  -7.000  -0.4480   0.08640   0.08202  -0.0209   1.0000   0.1558
-  -6.750  -0.4588   0.08412   0.07980  -0.0206   1.0000   0.1600
-  -6.500  -0.4847   0.07929   0.07495  -0.0287   1.0000   0.1662
-  -6.250  -0.4783   0.07718   0.07290  -0.0243   1.0000   0.1680
-  -6.000  -0.4744   0.07516   0.07091  -0.0219   1.0000   0.1710
-  -5.750  -0.4612   0.06965   0.06522  -0.0315   0.9966   0.1830
-  -5.500  -0.4309   0.06730   0.06289  -0.0316   0.9913   0.1898
-  -5.250  -0.3801   0.03672   0.02954  -0.0608   0.9884   0.1054
-  -5.000  -0.3390   0.03524   0.02804  -0.0644   0.9827   0.1094
-  -4.750  -0.2986   0.03253   0.02478  -0.0679   0.9764   0.1116
-  -4.500  -0.2547   0.03033   0.02199  -0.0716   0.9711   0.1152
-  -4.250  -0.2155   0.02874   0.02008  -0.0743   0.9639   0.1211
-  -4.000  -0.1732   0.02773   0.01892  -0.0774   0.9576   0.1286
-  -3.750  -0.1342   0.02656   0.01755  -0.0797   0.9500   0.1373
-  -3.500  -0.0916   0.02579   0.01667  -0.0827   0.9435   0.1516
-  -3.250  -0.0546   0.02511   0.01596  -0.0846   0.9347   0.1675
-  -3.000  -0.0135   0.02457   0.01552  -0.0873   0.9278   0.1881
-  -2.750   0.0233   0.02414   0.01507  -0.0891   0.9186   0.2119
-  -2.500   0.0621   0.02363   0.01470  -0.0912   0.9110   0.2389
-  -2.250   0.0998   0.02306   0.01425  -0.0931   0.9024   0.2721
-  -2.000   0.1378   0.02248   0.01382  -0.0950   0.8945   0.3153
-  -1.750   0.1759   0.02179   0.01342  -0.0968   0.8865   0.3683
-  -1.500   0.2137   0.02102   0.01305  -0.0985   0.8792   0.4420
-  -1.250   0.2473   0.02001   0.01285  -0.0990   0.8713   0.6046
-  -1.000   0.3011   0.01877   0.01237  -0.1014   0.8681   1.0000
-  -0.750   0.3304   0.01877   0.01214  -0.1016   0.8562   1.0000
-  -0.500   0.3765   0.01838   0.01153  -0.1044   0.8513   1.0000
-  -0.250   0.4028   0.01841   0.01142  -0.1039   0.8387   1.0000
-   0.000   0.4335   0.01835   0.01123  -0.1041   0.8283   1.0000
-   0.250   0.4720   0.01803   0.01077  -0.1054   0.8210   1.0000
-   0.500   0.4988   0.01808   0.01072  -0.1049   0.8090   1.0000
-   0.750   0.5348   0.01782   0.01035  -0.1058   0.8013   1.0000
-   1.000   0.5631   0.01782   0.01026  -0.1055   0.7900   1.0000
-   1.250   0.5902   0.01789   0.01027  -0.1051   0.7784   1.0000
-   1.500   0.6265   0.01763   0.00990  -0.1059   0.7710   1.0000
-   1.750   0.6506   0.01784   0.01006  -0.1050   0.7582   1.0000
-   2.000   0.6774   0.01798   0.01015  -0.1045   0.7469   1.0000
-   2.250   0.7110   0.01786   0.00993  -0.1050   0.7386   1.0000
-   2.500   0.7347   0.01814   0.01020  -0.1041   0.7259   1.0000
-   2.750   0.7611   0.01836   0.01039  -0.1036   0.7150   1.0000
-   3.000   0.7929   0.01834   0.01028  -0.1038   0.7060   1.0000
-   3.250   0.8161   0.01870   0.01065  -0.1029   0.6937   1.0000
-   3.500   0.8423   0.01895   0.01090  -0.1024   0.6829   1.0000
-   3.750   0.8728   0.01902   0.01091  -0.1024   0.6736   1.0000
-   4.000   0.8957   0.01941   0.01134  -0.1015   0.6614   1.0000
-   4.250   0.9217   0.01970   0.01163  -0.1010   0.6508   1.0000
-   4.500   0.9513   0.01982   0.01170  -0.1009   0.6411   1.0000
-   4.750   0.9739   0.02025   0.01220  -0.0999   0.6291   1.0000
-   5.000   0.9995   0.02057   0.01254  -0.0994   0.6185   1.0000
-   5.250   1.0286   0.02074   0.01268  -0.0992   0.6086   1.0000
-   5.500   1.0508   0.02119   0.01322  -0.0982   0.5964   1.0000
-   5.750   1.0755   0.02148   0.01355  -0.0974   0.5844   1.0000
-   6.000   1.1016   0.02159   0.01364  -0.0966   0.5709   1.0000
-   6.250   1.1279   0.02168   0.01370  -0.0958   0.5569   1.0000
-   6.500   1.1534   0.02184   0.01386  -0.0950   0.5432   1.0000
-   6.750   1.1757   0.02207   0.01413  -0.0937   0.5281   1.0000
-   7.000   1.1987   0.02216   0.01423  -0.0924   0.5114   1.0000
-   7.250   1.2212   0.02223   0.01428  -0.0911   0.4939   1.0000
-   7.500   1.2409   0.02240   0.01449  -0.0893   0.4761   1.0000
-   7.750   1.2594   0.02262   0.01478  -0.0875   0.4582   1.0000
-   8.000   1.2769   0.02285   0.01510  -0.0854   0.4393   1.0000
-   8.250   1.2933   0.02308   0.01539  -0.0832   0.4192   1.0000
-   8.500   1.3086   0.02333   0.01564  -0.0809   0.3979   1.0000
-   8.750   1.3197   0.02372   0.01612  -0.0779   0.3730   1.0000
-   9.000   1.3281   0.02421   0.01659  -0.0746   0.3445   1.0000
-   9.250   1.3316   0.02495   0.01728  -0.0707   0.3104   1.0000
-   9.500   1.3276   0.02605   0.01816  -0.0657   0.2733   1.0000
-   9.750   1.3204   0.02765   0.01952  -0.0608   0.2346   1.0000
-  10.000   1.3127   0.02965   0.02123  -0.0564   0.2023   1.0000
-  10.250   1.3077   0.03176   0.02308  -0.0527   0.1788   1.0000
-  10.500   1.3074   0.03378   0.02497  -0.0497   0.1606   1.0000
-  10.750   1.3110   0.03571   0.02678  -0.0472   0.1474   1.0000
-  11.000   1.3189   0.03755   0.02847  -0.0451   0.1369   1.0000
-  11.250   1.3289   0.03922   0.03015  -0.0434   0.1279   1.0000
-  11.500   1.3439   0.04092   0.03184  -0.0420   0.1203   1.0000
-  11.750   1.3595   0.04250   0.03339  -0.0408   0.1136   1.0000
-  12.000   1.3787   0.04430   0.03522  -0.0399   0.1076   1.0000
-  12.250   1.3912   0.04604   0.03710  -0.0386   0.1025   1.0000
-  12.500   1.4202   0.04801   0.03893  -0.0387   0.0966   1.0000
-  12.750   1.4239   0.05002   0.04125  -0.0367   0.0932   1.0000
-  13.000   1.4319   0.05205   0.04343  -0.0351   0.0895   1.0000
-  13.250   1.4604   0.05447   0.04574  -0.0355   0.0845   1.0000
-  13.500   1.4540   0.05696   0.04860  -0.0329   0.0825   1.0000
-  13.750   1.4507   0.05963   0.05156  -0.0309   0.0802   1.0000
-  14.000   1.4512   0.06218   0.05428  -0.0294   0.0775   1.0000
-  14.250   1.4783   0.06519   0.05713  -0.0297   0.0734   1.0000
-  14.500   1.4597   0.06850   0.06082  -0.0274   0.0727   1.0000
-  14.750   1.4404   0.07234   0.06501  -0.0257   0.0720   1.0000
-  15.000   1.4201   0.07667   0.06967  -0.0247   0.0712   1.0000
-  15.250   1.3983   0.08150   0.07481  -0.0243   0.0706   1.0000
-  15.500   1.3742   0.08690   0.08051  -0.0247   0.0701   1.0000
-  15.750   1.3474   0.09302   0.08692  -0.0259   0.0699   1.0000
-  16.000   1.3171   0.10006   0.09424  -0.0282   0.0701   1.0000
-  16.250   1.2836   0.10833   0.10277  -0.0317   0.0706   1.0000
-  16.500   1.2473   0.11795   0.11263  -0.0367   0.0714   1.0000
-  16.750   1.2101   0.12886   0.12372  -0.0430   0.0723   1.0000
-  17.000   1.1753   0.14068   0.13565  -0.0500   0.0732   1.0000
