--- regression/scripts/Vehicles/Airfoils/NACA_4412_polar_Re_200000.txt
@@ -1,115 +1,0 @@
-  
-       XFOIL         Version 6.96
-  
- Calculated polar for: NACA 4412                                       
-  
- 1 1 Reynolds number fixed          Mach number fixed         
-  
- xtrf =   1.000 (top)        1.000 (bottom)  
- Mach =   0.000     Re =     0.200 e 6     Ncrit =   9.000
-  
-   alpha    CL        CD       CDp       CM     Top_Xtr  Bot_Xtr
-  ------ -------- --------- --------- -------- -------- --------
-  -8.500  -0.4088   0.08983   0.08647  -0.0344   1.0000   0.0813
-  -8.250  -0.4231   0.08831   0.08501  -0.0315   1.0000   0.0823
-  -8.000  -0.4442   0.08695   0.08373  -0.0282   1.0000   0.0831
-  -7.750  -0.4937   0.05313   0.04929  -0.0671   0.9865   0.0673
-  -7.500  -0.4712   0.04655   0.04243  -0.0720   0.9817   0.0641
-  -7.250  -0.4535   0.03731   0.03244  -0.0777   0.9755   0.0622
-  -7.000  -0.4231   0.03226   0.02667  -0.0816   0.9717   0.0637
-  -6.750  -0.3952   0.02910   0.02295  -0.0831   0.9655   0.0647
-  -6.500  -0.3582   0.02703   0.02031  -0.0857   0.9620   0.0660
-  -6.250  -0.3204   0.02462   0.01772  -0.0887   0.9600   0.0684
-  -6.000  -0.2930   0.02357   0.01656  -0.0890   0.9523   0.0702
-  -5.750  -0.2544   0.02238   0.01517  -0.0913   0.9489   0.0725
-  -5.500  -0.2129   0.02134   0.01390  -0.0940   0.9465   0.0756
-  -5.250  -0.1843   0.02032   0.01270  -0.0942   0.9393   0.0785
-  -5.000  -0.1466   0.01934   0.01174  -0.0963   0.9352   0.0824
-  -4.750  -0.1047   0.01855   0.01086  -0.0990   0.9326   0.0877
-  -4.500  -0.0618   0.01758   0.00989  -0.1020   0.9308   0.0951
-  -4.250  -0.0351   0.01719   0.00943  -0.1016   0.9217   0.1030
-  -4.000   0.0030   0.01638   0.00871  -0.1036   0.9180   0.1150
-  -3.750   0.0426   0.01571   0.00806  -0.1058   0.9151   0.1299
-  -3.500   0.0687   0.01531   0.00771  -0.1053   0.9057   0.1437
-  -3.250   0.1047   0.01477   0.00720  -0.1066   0.9011   0.1620
-  -3.000   0.1342   0.01437   0.00686  -0.1068   0.8934   0.1821
-  -2.750   0.1659   0.01392   0.00653  -0.1073   0.8865   0.2077
-  -2.500   0.1964   0.01352   0.00623  -0.1075   0.8792   0.2377
-  -2.250   0.2255   0.01312   0.00593  -0.1075   0.8706   0.2707
-  -2.000   0.2541   0.01274   0.00566  -0.1074   0.8618   0.3065
-  -1.750   0.2836   0.01233   0.00537  -0.1074   0.8533   0.3489
-  -1.500   0.3098   0.01195   0.00520  -0.1069   0.8424   0.4005
-  -1.250   0.3373   0.01136   0.00499  -0.1066   0.8336   0.5000
-  -1.000   0.3583   0.01070   0.00498  -0.1046   0.8222   0.6832
-  -0.500   0.4431   0.00999   0.00468  -0.1085   0.8021   1.0000
-  -0.250   0.4683   0.01000   0.00456  -0.1077   0.7890   1.0000
-   0.000   0.4938   0.01005   0.00448  -0.1071   0.7758   1.0000
-   0.250   0.5198   0.01011   0.00441  -0.1065   0.7629   1.0000
-   0.500   0.5464   0.01019   0.00435  -0.1060   0.7504   1.0000
-   0.750   0.5733   0.01027   0.00429  -0.1056   0.7383   1.0000
-   1.000   0.5988   0.01039   0.00431  -0.1050   0.7246   1.0000
-   1.250   0.6247   0.01052   0.00435  -0.1045   0.7115   1.0000
-   1.500   0.6510   0.01066   0.00439  -0.1040   0.6992   1.0000
-   1.750   0.6778   0.01082   0.00443  -0.1036   0.6876   1.0000
-   2.000   0.7038   0.01097   0.00451  -0.1032   0.6752   1.0000
-   2.250   0.7297   0.01115   0.00463  -0.1027   0.6629   1.0000
-   2.500   0.7561   0.01133   0.00474  -0.1023   0.6517   1.0000
-   2.750   0.7828   0.01152   0.00483  -0.1020   0.6409   1.0000
-   3.000   0.8084   0.01171   0.00500  -0.1014   0.6289   1.0000
-   3.250   0.8345   0.01191   0.00515  -0.1010   0.6180   1.0000
-   3.500   0.8613   0.01213   0.00528  -0.1007   0.6079   1.0000
-   3.750   0.8865   0.01233   0.00549  -0.1002   0.5965   1.0000
-   4.000   0.9127   0.01256   0.00569  -0.0998   0.5863   1.0000
-   4.250   0.9390   0.01280   0.00587  -0.0995   0.5765   1.0000
-   4.500   0.9643   0.01303   0.00613  -0.0990   0.5659   1.0000
-   4.750   0.9905   0.01330   0.00636  -0.0986   0.5563   1.0000
-   5.000   1.0156   0.01352   0.00657  -0.0980   0.5448   1.0000
-   5.250   1.0396   0.01373   0.00679  -0.0972   0.5317   1.0000
-   5.500   1.0639   0.01396   0.00701  -0.0965   0.5191   1.0000
-   5.750   1.0888   0.01422   0.00724  -0.0959   0.5080   1.0000
-   6.000   1.1122   0.01444   0.00743  -0.0950   0.4940   1.0000
-   6.250   1.1342   0.01464   0.00764  -0.0938   0.4781   1.0000
-   6.500   1.1562   0.01486   0.00788  -0.0927   0.4627   1.0000
-   6.750   1.1783   0.01510   0.00815  -0.0916   0.4482   1.0000
-   7.000   1.2002   0.01536   0.00845  -0.0905   0.4333   1.0000
-   7.250   1.2215   0.01565   0.00876  -0.0893   0.4177   1.0000
-   7.500   1.2420   0.01596   0.00908  -0.0880   0.4012   1.0000
-   7.750   1.2616   0.01630   0.00943  -0.0866   0.3835   1.0000
-   8.000   1.2799   0.01670   0.00982  -0.0849   0.3635   1.0000
-   8.250   1.2967   0.01714   0.01026  -0.0831   0.3392   1.0000
-   8.500   1.3110   0.01772   0.01076  -0.0809   0.3110   1.0000
-   8.750   1.3218   0.01847   0.01138  -0.0781   0.2773   1.0000
-   9.000   1.3267   0.01944   0.01217  -0.0745   0.2406   1.0000
-   9.250   1.3283   0.02069   0.01320  -0.0706   0.1997   1.0000
-   9.500   1.3275   0.02223   0.01449  -0.0667   0.1617   1.0000
-   9.750   1.3266   0.02391   0.01597  -0.0630   0.1356   1.0000
-  10.000   1.3275   0.02555   0.01752  -0.0598   0.1187   1.0000
-  10.250   1.3289   0.02726   0.01916  -0.0569   0.1080   1.0000
-  10.500   1.3326   0.02887   0.02077  -0.0544   0.1001   1.0000
-  10.750   1.3366   0.03055   0.02249  -0.0520   0.0942   1.0000
-  11.000   1.3429   0.03209   0.02408  -0.0501   0.0890   1.0000
-  11.250   1.3438   0.03415   0.02607  -0.0479   0.0848   1.0000
-  11.500   1.3530   0.03558   0.02764  -0.0464   0.0812   1.0000
-  11.750   1.3601   0.03722   0.02934  -0.0450   0.0776   1.0000
-  12.000   1.3641   0.03917   0.03125  -0.0433   0.0746   1.0000
-  12.250   1.3713   0.04092   0.03309  -0.0418   0.0719   1.0000
-  12.500   1.3790   0.04262   0.03491  -0.0407   0.0690   1.0000
-  12.750   1.3857   0.04441   0.03676  -0.0395   0.0664   1.0000
-  13.000   1.3923   0.04629   0.03859  -0.0382   0.0639   1.0000
-  13.250   1.4001   0.04814   0.04054  -0.0369   0.0616   1.0000
-  13.500   1.4059   0.05012   0.04268  -0.0360   0.0593   1.0000
-  13.750   1.4115   0.05212   0.04477  -0.0351   0.0571   1.0000
-  14.000   1.4180   0.05407   0.04670  -0.0342   0.0549   1.0000
-  14.250   1.4266   0.05607   0.04875  -0.0329   0.0527   1.0000
-  14.500   1.4286   0.05852   0.05141  -0.0322   0.0509   1.0000
-  14.750   1.4308   0.06101   0.05404  -0.0316   0.0491   1.0000
-  15.000   1.4342   0.06336   0.05644  -0.0312   0.0473   1.0000
-  15.250   1.4467   0.06514   0.05814  -0.0297   0.0450   1.0000
-  15.500   1.4416   0.06845   0.06172  -0.0297   0.0440   1.0000
-  15.750   1.4385   0.07175   0.06524  -0.0296   0.0427   1.0000
-  16.000   1.4362   0.07501   0.06865  -0.0296   0.0414   1.0000
-  16.250   1.4351   0.07813   0.07188  -0.0298   0.0401   1.0000
-  16.500   1.4392   0.08059   0.07434  -0.0296   0.0388   1.0000
-  16.750   1.4401   0.08377   0.07759  -0.0291   0.0376   1.0000
-  17.000   1.4274   0.08864   0.08274  -0.0304   0.0369   1.0000
-  17.250   1.4151   0.09369   0.08804  -0.0318   0.0363   1.0000
