--- regression/scripts/Vehicles/Airfoils/NACA_4412_polar_Re_50000.txt
@@ -1,106 +1,0 @@
-  
-       XFOIL         Version 6.96
-  
- Calculated polar for: NACA 4412                                       
-  
- 1 1 Reynolds number fixed          Mach number fixed         
-  
- xtrf =   1.000 (top)        1.000 (bottom)  
- Mach =   0.000     Re =     0.050 e 6     Ncrit =   9.000
-  
-   alpha    CL        CD       CDp       CM     Top_Xtr  Bot_Xtr
-  ------ -------- --------- --------- -------- -------- --------
-  -9.250  -0.3484   0.11519   0.10799  -0.0307   1.0000   0.2459
-  -9.000  -0.3263   0.10991   0.10270  -0.0293   1.0000   0.2536
-  -8.750  -0.3430   0.10940   0.10231  -0.0284   1.0000   0.2620
-  -8.500  -0.3340   0.10549   0.09844  -0.0271   1.0000   0.2680
-  -8.250  -0.3380   0.10368   0.09671  -0.0255   1.0000   0.2771
-  -8.000  -0.3862   0.10566   0.09896  -0.0226   1.0000   0.2805
-  -7.750  -0.3424   0.09889   0.09208  -0.0219   1.0000   0.2911
-  -7.500  -0.3860   0.10017   0.09361  -0.0180   1.0000   0.2969
-  -7.250  -0.3576   0.09514   0.08853  -0.0169   1.0000   0.3077
-  -7.000  -0.3988   0.09590   0.08951  -0.0124   1.0000   0.3135
-  -6.750  -0.3812   0.09202   0.08562  -0.0104   1.0000   0.3258
-  -6.500  -0.4296   0.09289   0.08673  -0.0077   1.0000   0.3315
-  -6.250  -0.4111   0.08914   0.08297  -0.0042   1.0000   0.3457
-  -6.000  -0.4133   0.08662   0.08052  -0.0015   1.0000   0.3557
-  -5.750  -0.4339   0.08523   0.07926   0.0005   1.0000   0.3681
-  -5.500  -0.4490   0.08389   0.07801   0.0021   1.0000   0.3833
-  -5.250  -0.4372   0.08080   0.07495   0.0060   1.0000   0.3957
-  -5.000  -0.4394   0.07850   0.07271   0.0087   1.0000   0.4103
-  -4.750  -0.4076   0.05599   0.04887  -0.0396   1.0000   0.1988
-  -4.500  -0.3867   0.05206   0.04465  -0.0418   1.0000   0.1939
-  -4.250  -0.3630   0.04815   0.04039  -0.0444   1.0000   0.1911
-  -4.000  -0.3354   0.04419   0.03592  -0.0475   1.0000   0.1881
-  -3.750  -0.3092   0.04146   0.03274  -0.0493   1.0000   0.1908
-  -3.500  -0.2800   0.03898   0.02952  -0.0515   1.0000   0.1969
-  -3.250  -0.2589   0.03764   0.02820  -0.0515   1.0000   0.2030
-  -3.000  -0.2321   0.03627   0.02629  -0.0526   1.0000   0.2137
-  -2.750  -0.2112   0.03545   0.02553  -0.0525   1.0000   0.2250
-  -2.500  -0.1884   0.03470   0.02464  -0.0528   1.0000   0.2394
-  -2.250  -0.1656   0.03410   0.02391  -0.0531   1.0000   0.2564
-  -2.000  -0.1432   0.03369   0.02338  -0.0533   1.0000   0.2769
-  -1.750  -0.1176   0.03335   0.02302  -0.0542   0.9988   0.3012
-  -1.500  -0.0654   0.03317   0.02283  -0.0594   0.9883   0.3463
-  -1.250  -0.0158   0.03295   0.02274  -0.0639   0.9776   0.4048
-  -1.000   0.0297   0.03258   0.02279  -0.0676   0.9671   0.4910
-  -0.750   0.0671   0.03154   0.02306  -0.0683   0.9582   0.6968
-  -0.500   0.1031   0.03143   0.02286  -0.0702   0.9435   1.0000
-  -0.250   0.1469   0.03237   0.02333  -0.0742   0.9312   1.0000
-   0.000   0.1804   0.03318   0.02384  -0.0763   0.9178   1.0000
-   0.250   0.2127   0.03403   0.02444  -0.0782   0.9051   1.0000
-   0.500   0.2476   0.03490   0.02509  -0.0804   0.8929   1.0000
-   0.750   0.2900   0.03575   0.02574  -0.0836   0.8821   1.0000
-   1.000   0.3166   0.03657   0.02643  -0.0844   0.8694   1.0000
-   1.250   0.3420   0.03746   0.02721  -0.0850   0.8570   1.0000
-   1.500   0.3718   0.03836   0.02800  -0.0862   0.8458   1.0000
-   1.750   0.4130   0.03909   0.02862  -0.0888   0.8357   1.0000
-   2.000   0.4299   0.04014   0.02962  -0.0881   0.8232   1.0000
-   2.250   0.4529   0.04117   0.03059  -0.0883   0.8118   1.0000
-   2.500   0.4924   0.04188   0.03125  -0.0905   0.8022   1.0000
-   2.750   0.5095   0.04305   0.03239  -0.0899   0.7903   1.0000
-   3.000   0.5279   0.04426   0.03358  -0.0894   0.7788   1.0000
-   3.250   0.5662   0.04496   0.03427  -0.0912   0.7696   1.0000
-   3.500   0.5815   0.04627   0.03560  -0.0904   0.7577   1.0000
-   3.750   0.5962   0.04773   0.03705  -0.0896   0.7465   1.0000
-   4.000   0.6350   0.04832   0.03768  -0.0911   0.7370   1.0000
-   4.250   0.6464   0.04992   0.03930  -0.0901   0.7251   1.0000
-   4.500   0.6592   0.05153   0.04095  -0.0891   0.7135   1.0000
-   4.750   0.6949   0.05221   0.04167  -0.0901   0.7038   1.0000
-   5.000   0.7094   0.05375   0.04328  -0.0893   0.6918   1.0000
-   5.250   0.7169   0.05575   0.04532  -0.0880   0.6799   1.0000
-   5.500   0.7421   0.05688   0.04651  -0.0879   0.6689   1.0000
-   5.750   0.7723   0.05768   0.04742  -0.0881   0.6579   1.0000
-   6.000   0.7737   0.06012   0.04991  -0.0865   0.6450   1.0000
-   6.250   0.7860   0.06203   0.05189  -0.0857   0.6330   1.0000
-   6.500   0.8168   0.06274   0.05272  -0.0856   0.6214   1.0000
-   6.750   0.8415   0.06375   0.05385  -0.0851   0.6091   1.0000
-   7.000   0.8430   0.06634   0.05651  -0.0836   0.5952   1.0000
-   7.250   0.8513   0.06850   0.05876  -0.0825   0.5813   1.0000
-   7.500   0.8671   0.07004   0.06040  -0.0814   0.5669   1.0000
-   7.750   0.8865   0.07116   0.06166  -0.0802   0.5517   1.0000
-   8.000   0.9081   0.07193   0.06256  -0.0789   0.5355   1.0000
-   8.250   0.9309   0.07238   0.06316  -0.0773   0.5186   1.0000
-   8.500   0.9547   0.07259   0.06354  -0.0755   0.5015   1.0000
-   8.750   0.9857   0.07176   0.06289  -0.0734   0.4840   1.0000
-   9.000   1.2725   0.04155   0.03319  -0.0743   0.4435   1.0000
-   9.250   1.2985   0.04001   0.03163  -0.0714   0.4108   1.0000
-   9.500   1.3102   0.03953   0.03112  -0.0674   0.3762   1.0000
-   9.750   1.3184   0.03942   0.03081  -0.0631   0.3384   1.0000
-  10.000   1.3248   0.03991   0.03091  -0.0590   0.3012   1.0000
-  10.250   1.3265   0.04130   0.03202  -0.0549   0.2702   1.0000
-  10.500   1.3283   0.04312   0.03369  -0.0514   0.2452   1.0000
-  10.750   1.3401   0.04512   0.03548  -0.0492   0.2231   1.0000
-  11.000   1.3527   0.04742   0.03778  -0.0474   0.2062   1.0000
-  11.250   1.3686   0.04976   0.04008  -0.0460   0.1920   1.0000
-  11.500   1.3924   0.05231   0.04258  -0.0457   0.1796   1.0000
-  11.750   1.3929   0.05510   0.04569  -0.0429   0.1727   1.0000
-  12.000   1.4092   0.05796   0.04860  -0.0420   0.1644   1.0000
-  12.250   1.3997   0.06112   0.05211  -0.0387   0.1604   1.0000
-  12.500   1.4298   0.06425   0.05517  -0.0393   0.1521   1.0000
-  12.750   1.4065   0.06782   0.05914  -0.0353   0.1508   1.0000
-  13.000   1.3810   0.07194   0.06362  -0.0321   0.1498   1.0000
-  13.250   1.3520   0.07677   0.06876  -0.0299   0.1493   1.0000
-  13.500   1.3181   0.08256   0.07484  -0.0288   0.1496   1.0000
-  13.750   1.2801   0.08958   0.08211  -0.0292   0.1505   1.0000
-  14.000   1.2403   0.09792   0.09061  -0.0311   0.1517   1.0000
