+++ regression/scripts/Vehicles/Airfoils/NACA_63_412.txt
@@ -1,0 +1,56 @@
+NACA 63-412 AIRFOIL
+       26.       26.
+
+  0.000000  0.000000
+  0.003360  0.010710
+  0.005670  0.013200
+  0.010410  0.017190
+  0.022570  0.024600
+  0.047270  0.035440
+  0.072180  0.043790
+  0.097180  0.050630
+  0.147350  0.061380
+  0.197650  0.069290
+  0.248000  0.074990
+  0.298400  0.078720
+  0.348820  0.080590
+  0.399240  0.080620
+  0.449640  0.078940
+  0.500000  0.075670
+  0.550310  0.071250
+  0.600570  0.065620
+  0.650760  0.058990
+  0.700870  0.051530
+  0.750890  0.043440
+  0.800840  0.034920
+  0.850700  0.026180
+  0.900490  0.017390
+  0.950230  0.008810
+  1.000000  0.000000
+
+  0.000000  0.000000
+  0.006640 -0.008710
+  0.009330 -0.010400
+  0.014590 -0.012910
+  0.027430 -0.017160
+  0.052730 -0.022800
+  0.077820 -0.026850
+  0.102820 -0.029950
+  0.152650 -0.034460
+  0.202350 -0.037450
+  0.252000 -0.039190
+  0.301600 -0.039840
+  0.351118 -0.039390
+  0.400760 -0.037780
+  0.450350 -0.035140
+  0.500000 -0.031640
+  0.549690 -0.027450
+  0.599430 -0.022780
+  0.649240 -0.017990
+  0.699130 -0.012650
+  0.749110 -0.007640
+  0.799160 -0.003080
+  0.849300  0.000740
+  0.899510  0.003290
+  0.949770  0.003300
+  1.000000  0.000000
