+++ regression/scripts/Vehicles/Airfoils/Polars/NACA_63_412_polar_Re_100000.txt
@@ -1,0 +1,114 @@
+  
+       XFOIL         Version 6.96
+  
+ Calculated polar for: NACA 63-412 AIRFOIL                             
+  
+ 1 1 Reynolds number fixed          Mach number fixed         
+  
+ xtrf =   1.000 (top)        1.000 (bottom)  
+ Mach =   0.000     Re =     0.100 e 6     Ncrit =   5.000
+  
+   alpha    CL        CD       CDp       CM     Top_Xtr  Bot_Xtr
+  ------ -------- --------- --------- -------- -------- --------
+ -10.500  -0.4783   0.08951   0.08431  -0.0435   1.0000   0.0331
+ -10.250  -0.4905   0.08189   0.07675  -0.0480   1.0000   0.0327
+ -10.000  -0.5183   0.07077   0.06565  -0.0563   1.0000   0.0317
+  -9.750  -0.5512   0.06300   0.05780  -0.0619   1.0000   0.0311
+  -9.500  -0.5841   0.05816   0.05286  -0.0629   1.0000   0.0306
+  -9.250  -0.6054   0.05429   0.04882  -0.0623   1.0000   0.0305
+  -9.000  -0.6197   0.05077   0.04510  -0.0608   1.0000   0.0305
+  -8.750  -0.6276   0.04780   0.04192  -0.0589   1.0000   0.0306
+  -8.500  -0.6327   0.04510   0.03900  -0.0567   1.0000   0.0309
+  -8.250  -0.6350   0.04262   0.03627  -0.0543   1.0000   0.0312
+  -8.000  -0.6320   0.04016   0.03353  -0.0525   0.9994   0.0317
+  -7.750  -0.6042   0.03712   0.03003  -0.0553   0.9922   0.0335
+  -7.500  -0.5747   0.03401   0.02632  -0.0577   0.9856   0.0354
+  -7.250  -0.5451   0.03138   0.02311  -0.0594   0.9786   0.0365
+  -7.000  -0.5139   0.02893   0.02042  -0.0612   0.9732   0.0378
+  -6.750  -0.4828   0.02752   0.01891  -0.0629   0.9664   0.0402
+  -6.500  -0.4482   0.02608   0.01724  -0.0648   0.9617   0.0430
+  -6.250  -0.4174   0.02465   0.01560  -0.0656   0.9546   0.0449
+  -6.000  -0.3825   0.02324   0.01405  -0.0673   0.9499   0.0474
+  -5.750  -0.3523   0.02220   0.01305  -0.0684   0.9426   0.0511
+  -5.500  -0.3195   0.02124   0.01200  -0.0697   0.9363   0.0551
+  -5.250  -0.2886   0.02028   0.01096  -0.0706   0.9293   0.0590
+  -5.000  -0.2581   0.01951   0.01018  -0.0716   0.9217   0.0661
+  -4.750  -0.2267   0.01870   0.00933  -0.0727   0.9150   0.0745
+  -4.500  -0.1977   0.01799   0.00862  -0.0734   0.9068   0.0891
+  -4.250  -0.1667   0.01711   0.00789  -0.0745   0.9003   0.1256
+  -4.000  -0.1415   0.01578   0.00726  -0.0752   0.8916   0.2573
+  -3.750  -0.1144   0.01487   0.00712  -0.0757   0.8849   0.4407
+  -3.500  -0.0882   0.01474   0.00715  -0.0751   0.8763   0.5184
+  -3.250  -0.0597   0.01473   0.00720  -0.0748   0.8696   0.5748
+  -3.000  -0.0344   0.01484   0.00734  -0.0738   0.8610   0.6176
+  -2.750  -0.0073   0.01497   0.00745  -0.0729   0.8543   0.6488
+  -2.500   0.0176   0.01509   0.00754  -0.0718   0.8457   0.6695
+  -2.250   0.0463   0.01510   0.00744  -0.0716   0.8393   0.6843
+  -2.000   0.0725   0.01510   0.00736  -0.0712   0.8305   0.6966
+  -1.750   0.1018   0.01505   0.00720  -0.0712   0.8244   0.7083
+  -1.500   0.1269   0.01507   0.00718  -0.0705   0.8155   0.7176
+  -1.250   0.1558   0.01502   0.00703  -0.0704   0.8093   0.7277
+  -1.000   0.1813   0.01504   0.00701  -0.0700   0.8003   0.7376
+  -0.750   0.2092   0.01500   0.00692  -0.0696   0.7941   0.7465
+  -0.500   0.2351   0.01503   0.00691  -0.0693   0.7852   0.7568
+  -0.250   0.2622   0.01500   0.00685  -0.0688   0.7788   0.7656
+   0.000   0.2876   0.01504   0.00689  -0.0683   0.7701   0.7753
+   0.250   0.3154   0.01503   0.00683  -0.0681   0.7636   0.7855
+   0.500   0.3399   0.01508   0.00690  -0.0674   0.7551   0.7947
+   0.750   0.3672   0.01508   0.00687  -0.0671   0.7484   0.8049
+   1.000   0.3923   0.01514   0.00697  -0.0665   0.7400   0.8154
+   1.250   0.4185   0.01514   0.00697  -0.0659   0.7333   0.8252
+   1.500   0.4434   0.01521   0.00708  -0.0653   0.7250   0.8362
+   1.750   0.4699   0.01523   0.00711  -0.0648   0.7182   0.8476
+   2.000   0.4939   0.01529   0.00723  -0.0640   0.7099   0.8588
+   2.250   0.5195   0.01529   0.00727  -0.0633   0.7029   0.8706
+   2.500   0.5438   0.01535   0.00740  -0.0626   0.6946   0.8834
+   2.750   0.5695   0.01534   0.00745  -0.0619   0.6874   0.8970
+   3.000   0.5940   0.01539   0.00759  -0.0612   0.6786   0.9119
+   3.250   0.6222   0.01536   0.00761  -0.0611   0.6713   0.9278
+   3.500   0.6518   0.01542   0.00781  -0.0615   0.6616   0.9469
+   3.750   0.6880   0.01543   0.00790  -0.0632   0.6529   0.9711
+   4.000   0.7192   0.01553   0.00807  -0.0642   0.6423   1.0000
+   4.250   0.7467   0.01568   0.00829  -0.0643   0.6301   1.0000
+   4.500   0.7742   0.01578   0.00844  -0.0643   0.6163   1.0000
+   4.750   0.8012   0.01585   0.00855  -0.0642   0.5995   1.0000
+   5.000   0.8279   0.01590   0.00863  -0.0638   0.5803   1.0000
+   5.250   0.8527   0.01600   0.00877  -0.0632   0.5553   1.0000
+   5.500   0.8763   0.01611   0.00883  -0.0623   0.5224   1.0000
+   6.000   0.9189   0.01676   0.00924  -0.0599   0.4276   1.0000
+   6.250   0.9359   0.01747   0.00961  -0.0583   0.3536   1.0000
+   6.500   0.9469   0.01877   0.01034  -0.0561   0.2602   1.0000
+   6.750   0.9566   0.02034   0.01138  -0.0541   0.1783   1.0000
+   7.000   0.9673   0.02189   0.01254  -0.0523   0.1203   1.0000
+   7.250   0.9796   0.02327   0.01370  -0.0506   0.0900   1.0000
+   7.500   0.9931   0.02449   0.01485  -0.0489   0.0736   1.0000
+   7.750   1.0056   0.02574   0.01605  -0.0472   0.0638   1.0000
+   8.000   1.0186   0.02687   0.01724  -0.0455   0.0568   1.0000
+   8.250   1.0281   0.02813   0.01853  -0.0434   0.0522   1.0000
+   8.500   1.0395   0.02934   0.01984  -0.0416   0.0482   1.0000
+   8.750   1.0494   0.03070   0.02122  -0.0398   0.0451   1.0000
+   9.000   1.0595   0.03220   0.02276  -0.0381   0.0426   1.0000
+   9.250   1.0728   0.03354   0.02423  -0.0368   0.0397   1.0000
+   9.500   1.0861   0.03501   0.02578  -0.0355   0.0378   1.0000
+   9.750   1.0998   0.03659   0.02741  -0.0344   0.0363   1.0000
+  10.000   1.1158   0.03843   0.02925  -0.0335   0.0349   1.0000
+  10.250   1.1351   0.04023   0.03120  -0.0327   0.0337   1.0000
+  10.500   1.1522   0.04202   0.03321  -0.0319   0.0322   1.0000
+  10.750   1.1665   0.04388   0.03527  -0.0310   0.0308   1.0000
+  11.000   1.1787   0.04583   0.03742  -0.0300   0.0296   1.0000
+  11.250   1.1908   0.04801   0.03978  -0.0291   0.0288   1.0000
+  11.500   1.2009   0.05036   0.04230  -0.0282   0.0282   1.0000
+  11.750   1.2091   0.05293   0.04503  -0.0273   0.0277   1.0000
+  12.000   1.2152   0.05586   0.04815  -0.0265   0.0273   1.0000
+  12.250   1.2175   0.05930   0.05178  -0.0256   0.0269   1.0000
+  12.500   1.2111   0.06277   0.05558  -0.0246   0.0268   1.0000
+  12.750   1.2019   0.06659   0.05972  -0.0241   0.0266   1.0000
+  13.000   1.1897   0.07081   0.06425  -0.0240   0.0265   1.0000
+  13.250   1.1750   0.07547   0.06920  -0.0245   0.0265   1.0000
+  13.500   1.1579   0.08067   0.07468  -0.0257   0.0264   1.0000
+  13.750   1.1390   0.08644   0.08071  -0.0276   0.0264   1.0000
+  14.000   1.1186   0.09282   0.08733  -0.0304   0.0264   1.0000
+  14.250   1.0970   0.09992   0.09465  -0.0341   0.0264   1.0000
+  14.500   1.0747   0.10777   0.10271  -0.0387   0.0265   1.0000
+  14.750   1.0520   0.11657   0.11168  -0.0443   0.0267   1.0000
+  15.000   1.0291   0.12642   0.12167  -0.0508   0.0269   1.0000
