+++ regression/scripts/Vehicles/Airfoils/Polars/NACA_63_412_polar_Re_200000.txt
@@ -1,0 +1,123 @@
+  
+       XFOIL         Version 6.96
+  
+ Calculated polar for: NACA 63-412 AIRFOIL                             
+  
+ 1 1 Reynolds number fixed          Mach number fixed         
+  
+ xtrf =   1.000 (top)        1.000 (bottom)  
+ Mach =   0.000     Re =     0.200 e 6     Ncrit =   5.000
+  
+   alpha    CL        CD       CDp       CM     Top_Xtr  Bot_Xtr
+  ------ -------- --------- --------- -------- -------- --------
+ -11.250  -0.5922   0.06654   0.06268  -0.0567   1.0000   0.0176
+ -11.000  -0.6185   0.05869   0.05470  -0.0629   1.0000   0.0176
+ -10.750  -0.6387   0.05337   0.04925  -0.0664   1.0000   0.0175
+ -10.500  -0.6626   0.04883   0.04454  -0.0678   1.0000   0.0175
+ -10.250  -0.6859   0.04557   0.04111  -0.0665   1.0000   0.0175
+ -10.000  -0.7050   0.04177   0.03699  -0.0647   1.0000   0.0176
+  -9.750  -0.7158   0.03843   0.03329  -0.0625   1.0000   0.0178
+  -9.500  -0.7160   0.03625   0.03096  -0.0604   1.0000   0.0180
+  -9.250  -0.7064   0.03457   0.02915  -0.0595   0.9983   0.0184
+  -9.000  -0.6795   0.03247   0.02681  -0.0618   0.9912   0.0190
+  -8.750  -0.6503   0.03076   0.02489  -0.0642   0.9852   0.0201
+  -8.500  -0.6232   0.02850   0.02223  -0.0659   0.9777   0.0213
+  -8.250  -0.5942   0.02623   0.01947  -0.0675   0.9714   0.0224
+  -8.000  -0.5650   0.02427   0.01728  -0.0689   0.9649   0.0232
+  -7.750  -0.5349   0.02305   0.01598  -0.0703   0.9581   0.0241
+  -7.500  -0.5035   0.02210   0.01492  -0.0718   0.9514   0.0255
+  -7.250  -0.4725   0.02108   0.01370  -0.0729   0.9445   0.0272
+  -7.000  -0.4418   0.02000   0.01244  -0.0739   0.9370   0.0284
+  -6.750  -0.4122   0.01882   0.01117  -0.0747   0.9295   0.0295
+  -6.500  -0.3830   0.01800   0.01034  -0.0755   0.9209   0.0313
+  -6.250  -0.3545   0.01734   0.00960  -0.0760   0.9118   0.0334
+  -6.000  -0.3259   0.01664   0.00880  -0.0764   0.9033   0.0353
+  -5.750  -0.2995   0.01603   0.00809  -0.0763   0.8936   0.0369
+  -5.500  -0.2731   0.01528   0.00731  -0.0765   0.8852   0.0397
+  -5.250  -0.2468   0.01483   0.00680  -0.0765   0.8759   0.0433
+  -5.000  -0.2200   0.01439   0.00626  -0.0764   0.8677   0.0467
+  -4.750  -0.1937   0.01387   0.00571  -0.0764   0.8592   0.0521
+  -4.500  -0.1670   0.01347   0.00526  -0.0764   0.8511   0.0603
+  -4.250  -0.1402   0.01303   0.00483  -0.0764   0.8430   0.0753
+  -4.000  -0.1138   0.01252   0.00446  -0.0765   0.8350   0.1122
+  -3.750  -0.0878   0.01177   0.00405  -0.0768   0.8273   0.2014
+  -3.500  -0.0625   0.01089   0.00375  -0.0771   0.8195   0.3463
+  -3.250  -0.0360   0.01049   0.00365  -0.0771   0.8119   0.4466
+  -3.000  -0.0089   0.01033   0.00360  -0.0769   0.8044   0.5004
+  -2.750   0.0183   0.01025   0.00359  -0.0767   0.7970   0.5456
+  -2.500   0.0455   0.01022   0.00360  -0.0764   0.7897   0.5836
+  -2.250   0.0725   0.01020   0.00360  -0.0760   0.7820   0.6103
+  -2.000   0.1000   0.01022   0.00359  -0.0757   0.7749   0.6315
+  -1.750   0.1275   0.01022   0.00357  -0.0755   0.7672   0.6441
+  -1.500   0.1553   0.01023   0.00353  -0.0754   0.7602   0.6549
+  -1.250   0.1832   0.01024   0.00349  -0.0753   0.7523   0.6653
+  -1.000   0.2112   0.01025   0.00346  -0.0752   0.7455   0.6748
+  -0.750   0.2389   0.01026   0.00345  -0.0751   0.7375   0.6836
+  -0.500   0.2672   0.01029   0.00341  -0.0751   0.7308   0.6934
+  -0.250   0.2946   0.01030   0.00343  -0.0750   0.7227   0.7012
+   0.000   0.3227   0.01034   0.00341  -0.0749   0.7161   0.7103
+   0.250   0.3501   0.01036   0.00346  -0.0748   0.7080   0.7184
+   0.500   0.3780   0.01040   0.00347  -0.0747   0.7013   0.7272
+   0.750   0.4054   0.01043   0.00354  -0.0746   0.6932   0.7357
+   1.000   0.4330   0.01048   0.00357  -0.0744   0.6864   0.7441
+   1.250   0.4605   0.01053   0.00366  -0.0743   0.6783   0.7531
+   1.500   0.4877   0.01058   0.00371  -0.0741   0.6714   0.7613
+   1.750   0.5152   0.01064   0.00382  -0.0740   0.6632   0.7707
+   2.000   0.5420   0.01070   0.00390  -0.0736   0.6560   0.7789
+   2.250   0.5693   0.01077   0.00402  -0.0735   0.6475   0.7882
+   2.500   0.5959   0.01084   0.00412  -0.0731   0.6400   0.7969
+   2.750   0.6226   0.01091   0.00426  -0.0728   0.6309   0.8060
+   3.000   0.6491   0.01100   0.00437  -0.0725   0.6226   0.8156
+   3.250   0.6749   0.01107   0.00452  -0.0720   0.6125   0.8246
+   3.500   0.7010   0.01115   0.00465  -0.0715   0.6008   0.8347
+   3.750   0.7261   0.01123   0.00477  -0.0708   0.5873   0.8448
+   4.000   0.7506   0.01131   0.00489  -0.0701   0.5709   0.8550
+   4.250   0.7749   0.01141   0.00500  -0.0692   0.5526   0.8663
+   4.750   0.8202   0.01171   0.00523  -0.0669   0.4968   0.8916
+   5.000   0.8410   0.01193   0.00537  -0.0655   0.4590   0.9069
+   5.250   0.8612   0.01220   0.00555  -0.0640   0.4164   0.9258
+   5.500   0.8827   0.01269   0.00582  -0.0631   0.3565   0.9584
+   5.750   0.9022   0.01365   0.00637  -0.0624   0.2750   1.0000
+   6.000   0.9187   0.01482   0.00710  -0.0613   0.1962   1.0000
+   6.250   0.9358   0.01594   0.00787  -0.0603   0.1347   1.0000
+   6.500   0.9536   0.01697   0.00863  -0.0593   0.0915   1.0000
+   6.750   0.9724   0.01788   0.00939  -0.0583   0.0661   1.0000
+   7.000   0.9919   0.01869   0.01013  -0.0574   0.0526   1.0000
+   7.250   1.0110   0.01948   0.01090  -0.0564   0.0446   1.0000
+   7.500   1.0306   0.02020   0.01168  -0.0555   0.0394   1.0000
+   7.750   1.0476   0.02108   0.01255  -0.0543   0.0353   1.0000
+   8.000   1.0644   0.02193   0.01349  -0.0529   0.0328   1.0000
+   8.250   1.0800   0.02277   0.01441  -0.0514   0.0306   1.0000
+   8.500   1.0930   0.02368   0.01537  -0.0496   0.0288   1.0000
+   8.750   1.1019   0.02488   0.01659  -0.0473   0.0272   1.0000
+   9.000   1.1143   0.02590   0.01770  -0.0456   0.0260   1.0000
+   9.250   1.1268   0.02695   0.01885  -0.0440   0.0246   1.0000
+   9.500   1.1387   0.02808   0.02006  -0.0424   0.0235   1.0000
+   9.750   1.1498   0.02932   0.02136  -0.0409   0.0227   1.0000
+  10.000   1.1604   0.03066   0.02276  -0.0395   0.0220   1.0000
+  10.250   1.1699   0.03218   0.02433  -0.0381   0.0214   1.0000
+  10.500   1.1798   0.03390   0.02609  -0.0367   0.0208   1.0000
+  10.750   1.1923   0.03538   0.02772  -0.0356   0.0204   1.0000
+  11.000   1.2047   0.03697   0.02945  -0.0345   0.0199   1.0000
+  11.250   1.2168   0.03866   0.03129  -0.0334   0.0195   1.0000
+  11.500   1.2281   0.04043   0.03321  -0.0324   0.0190   1.0000
+  11.750   1.2381   0.04228   0.03521  -0.0315   0.0185   1.0000
+  12.000   1.2465   0.04415   0.03723  -0.0306   0.0180   1.0000
+  12.250   1.2536   0.04614   0.03935  -0.0298   0.0176   1.0000
+  12.500   1.2594   0.04821   0.04155  -0.0292   0.0171   1.0000
+  12.750   1.2644   0.05049   0.04395  -0.0286   0.0168   1.0000
+  13.000   1.2684   0.05302   0.04661  -0.0281   0.0166   1.0000
+  13.250   1.2708   0.05589   0.04961  -0.0277   0.0164   1.0000
+  13.500   1.2700   0.05924   0.05315  -0.0274   0.0162   1.0000
+  13.750   1.2651   0.06291   0.05708  -0.0274   0.0160   1.0000
+  14.000   1.2577   0.06701   0.06145  -0.0279   0.0159   1.0000
+  14.250   1.2476   0.07160   0.06630  -0.0287   0.0158   1.0000
+  14.500   1.2353   0.07669   0.07165  -0.0301   0.0158   1.0000
+  14.750   1.2206   0.08236   0.07757  -0.0321   0.0157   1.0000
+  15.000   1.2039   0.08874   0.08420  -0.0349   0.0156   1.0000
+  15.250   1.1855   0.09582   0.09152  -0.0384   0.0156   1.0000
+  15.500   1.1654   0.10373   0.09966  -0.0428   0.0156   1.0000
+  15.750   1.1439   0.11253   0.10868  -0.0481   0.0156   1.0000
+  16.000   1.1203   0.12262   0.11898  -0.0546   0.0157   1.0000
+  16.250   1.0940   0.13428   0.13082  -0.0624   0.0158   1.0000
+  16.500   1.0626   0.14877   0.14548  -0.0720   0.0160   1.0000
