+++ regression/scripts/Vehicles/Airfoils/Polars/NACA_63_412_polar_Re_50000.txt
@@ -1,0 +1,109 @@
+  
+       XFOIL         Version 6.96
+  
+ Calculated polar for: NACA 63-412 AIRFOIL                             
+  
+ 1 1 Reynolds number fixed          Mach number fixed         
+  
+ xtrf =   1.000 (top)        1.000 (bottom)  
+ Mach =   0.000     Re =     0.050 e 6     Ncrit =   5.000
+  
+   alpha    CL        CD       CDp       CM     Top_Xtr  Bot_Xtr
+  ------ -------- --------- --------- -------- -------- --------
+ -10.750  -0.4759   0.10169   0.09432  -0.0425   1.0000   0.0575
+ -10.500  -0.4749   0.09720   0.08987  -0.0440   1.0000   0.0571
+ -10.250  -0.4768   0.09223   0.08495  -0.0462   1.0000   0.0567
+ -10.000  -0.4815   0.08675   0.07952  -0.0491   1.0000   0.0561
+  -9.750  -0.4911   0.08100   0.07383  -0.0525   1.0000   0.0556
+  -9.500  -0.5061   0.07551   0.06838  -0.0559   1.0000   0.0551
+  -9.250  -0.5259   0.07084   0.06374  -0.0581   1.0000   0.0546
+  -9.000  -0.5460   0.06689   0.05978  -0.0588   1.0000   0.0542
+  -8.750  -0.5620   0.06310   0.05590  -0.0587   1.0000   0.0539
+  -8.500  -0.5747   0.05950   0.05216  -0.0580   1.0000   0.0538
+  -8.250  -0.5831   0.05611   0.04857  -0.0568   1.0000   0.0538
+  -8.000  -0.5881   0.05292   0.04513  -0.0554   1.0000   0.0543
+  -7.750  -0.5903   0.04991   0.04178  -0.0538   1.0000   0.0552
+  -7.500  -0.5891   0.04703   0.03848  -0.0521   1.0000   0.0564
+  -7.250  -0.5841   0.04432   0.03525  -0.0505   1.0000   0.0575
+  -7.000  -0.5745   0.04174   0.03244  -0.0491   1.0000   0.0586
+  -6.750  -0.5624   0.03952   0.03001  -0.0478   1.0000   0.0596
+  -6.500  -0.5485   0.03754   0.02781  -0.0465   1.0000   0.0610
+  -6.250  -0.5334   0.03591   0.02598  -0.0453   1.0000   0.0636
+  -6.000  -0.5169   0.03434   0.02410  -0.0441   1.0000   0.0670
+  -5.750  -0.4988   0.03284   0.02222  -0.0428   1.0000   0.0695
+  -5.500  -0.4673   0.03111   0.02040  -0.0439   0.9948   0.0726
+  -5.250  -0.4338   0.02988   0.01905  -0.0454   0.9881   0.0790
+  -5.000  -0.4009   0.02870   0.01773  -0.0465   0.9815   0.0855
+  -4.750  -0.3671   0.02764   0.01657  -0.0480   0.9749   0.0939
+  -4.500  -0.3361   0.02661   0.01552  -0.0493   0.9669   0.1064
+  -4.250  -0.3029   0.02552   0.01439  -0.0511   0.9598   0.1257
+  -3.750  -0.2443   0.02196   0.01275  -0.0550   0.9451   0.4112
+  -3.500  -0.2174   0.02206   0.01317  -0.0540   0.9366   0.5528
+  -3.250  -0.1933   0.02254   0.01373  -0.0520   0.9280   0.6249
+  -3.000  -0.1707   0.02326   0.01446  -0.0490   0.9199   0.6833
+  -2.750  -0.1506   0.02372   0.01484  -0.0460   0.9111   0.7197
+  -2.500  -0.1219   0.02392   0.01488  -0.0452   0.9040   0.7447
+  -2.250  -0.0966   0.02395   0.01473  -0.0445   0.8955   0.7618
+  -2.000  -0.0643   0.02393   0.01452  -0.0452   0.8889   0.7773
+  -1.750  -0.0386   0.02390   0.01434  -0.0449   0.8802   0.7916
+  -1.500  -0.0070   0.02383   0.01411  -0.0456   0.8734   0.8054
+  -1.250   0.0176   0.02381   0.01397  -0.0452   0.8647   0.8192
+  -1.000   0.0489   0.02374   0.01377  -0.0458   0.8580   0.8326
+  -0.750   0.0728   0.02373   0.01368  -0.0453   0.8492   0.8460
+  -0.500   0.1040   0.02366   0.01352  -0.0459   0.8425   0.8594
+  -0.250   0.1285   0.02366   0.01346  -0.0455   0.8338   0.8738
+   0.000   0.1608   0.02358   0.01332  -0.0464   0.8271   0.8882
+   0.250   0.1881   0.02360   0.01331  -0.0466   0.8187   0.9038
+   0.500   0.2238   0.02355   0.01322  -0.0482   0.8120   0.9190
+   0.750   0.2591   0.02359   0.01325  -0.0501   0.8044   0.9352
+   1.000   0.3013   0.02360   0.01325  -0.0532   0.7974   0.9508
+   1.250   0.3476   0.02362   0.01328  -0.0572   0.7911   0.9662
+   1.500   0.3902   0.02373   0.01342  -0.0608   0.7828   0.9869
+   1.750   0.4180   0.02388   0.01356  -0.0616   0.7749   1.0000
+   2.000   0.4383   0.02413   0.01381  -0.0611   0.7659   1.0000
+   2.250   0.4599   0.02448   0.01416  -0.0609   0.7568   1.0000
+   2.500   0.4894   0.02467   0.01437  -0.0616   0.7495   1.0000
+   2.750   0.5119   0.02512   0.01485  -0.0616   0.7398   1.0000
+   3.000   0.5443   0.02527   0.01505  -0.0625   0.7332   1.0000
+   3.250   0.5662   0.02580   0.01564  -0.0623   0.7228   1.0000
+   3.500   0.6008   0.02586   0.01576  -0.0632   0.7167   1.0000
+   3.750   0.6216   0.02645   0.01646  -0.0628   0.7054   1.0000
+   4.000   0.6482   0.02682   0.01693  -0.0629   0.6962   1.0000
+   4.250   0.6784   0.02699   0.01721  -0.0631   0.6877   1.0000
+   4.500   0.7013   0.02748   0.01785  -0.0627   0.6762   1.0000
+   4.750   0.7294   0.02769   0.01820  -0.0626   0.6661   1.0000
+   5.000   0.7613   0.02762   0.01828  -0.0626   0.6560   1.0000
+   5.250   0.7860   0.02782   0.01865  -0.0619   0.6423   1.0000
+   5.500   0.8128   0.02777   0.01881  -0.0611   0.6275   1.0000
+   5.750   0.8424   0.02738   0.01859  -0.0602   0.6106   1.0000
+   6.000   0.8654   0.02716   0.01854  -0.0585   0.5884   1.0000
+   6.250   0.8910   0.02650   0.01802  -0.0565   0.5606   1.0000
+   6.500   0.9104   0.02608   0.01767  -0.0539   0.5234   1.0000
+   6.750   0.9270   0.02590   0.01748  -0.0511   0.4754   1.0000
+   7.000   0.9401   0.02612   0.01757  -0.0483   0.4122   1.0000
+   7.250   0.9485   0.02693   0.01791  -0.0451   0.3202   1.0000
+   7.500   0.9472   0.02883   0.01909  -0.0417   0.2324   1.0000
+   7.750   0.9432   0.03108   0.02083  -0.0386   0.1761   1.0000
+   8.000   0.9425   0.03336   0.02280  -0.0362   0.1413   1.0000
+   8.250   0.9449   0.03557   0.02483  -0.0343   0.1206   1.0000
+   8.500   0.9496   0.03769   0.02684  -0.0326   0.1057   1.0000
+   8.750   0.9578   0.03965   0.02879  -0.0312   0.0943   1.0000
+   9.000   0.9689   0.04152   0.03072  -0.0299   0.0862   1.0000
+   9.250   0.9817   0.04332   0.03250  -0.0288   0.0788   1.0000
+   9.500   1.0031   0.04488   0.03423  -0.0278   0.0722   1.0000
+   9.750   1.0260   0.04648   0.03585  -0.0271   0.0667   1.0000
+  10.000   1.0540   0.04828   0.03785  -0.0266   0.0615   1.0000
+  10.250   1.0829   0.05037   0.04020  -0.0264   0.0578   1.0000
+  10.500   1.1069   0.05267   0.04255  -0.0263   0.0549   1.0000
+  10.750   1.1236   0.05550   0.04568  -0.0257   0.0525   1.0000
+  11.000   1.1309   0.05848   0.04908  -0.0245   0.0507   1.0000
+  11.250   1.1352   0.06171   0.05266  -0.0235   0.0494   1.0000
+  11.500   1.1352   0.06522   0.05650  -0.0225   0.0487   1.0000
+  11.750   1.1305   0.06897   0.06057  -0.0216   0.0482   1.0000
+  12.000   1.1213   0.07308   0.06499  -0.0212   0.0478   1.0000
+  12.250   1.1079   0.07762   0.06981  -0.0213   0.0476   1.0000
+  12.500   1.0904   0.08277   0.07525  -0.0222   0.0476   1.0000
+  12.750   1.0688   0.08870   0.08144  -0.0240   0.0477   1.0000
+  13.000   1.0430   0.09570   0.08869  -0.0272   0.0480   1.0000
+  13.250   1.0136   0.10409   0.09729  -0.0319   0.0485   1.0000
+  13.500   0.9815   0.11425   0.10762  -0.0384   0.0493   1.0000
