+++ regression/scripts/Vehicles/X57_Maxwell_Mod2.py
@@ -1,0 +1,459 @@
+# Boeing_737.py
+#
+# Created: Feb 2020, M. Clarke
+#          Sep 2020, M. Clarke
+
+""" setup file for the X57-Maxwell Modification 2 Electric Aircraft
+"""
+
+# ----------------------------------------------------------------------
+#   Imports
+# ----------------------------------------------------------------------
+import SUAVE
+from SUAVE.Core import Units , Data 
+from SUAVE.Components.Energy.Networks.Battery_Propeller import Battery_Propeller
+from SUAVE.Methods.Propulsion                           import propeller_design
+from SUAVE.Methods.Power.Battery.Sizing                 import initialize_from_mass
+from SUAVE.Methods.Propulsion.electric_motor_sizing     import size_optimal_motor
+
+import numpy as np 
+from copy import deepcopy
+
+# ----------------------------------------------------------------------
+#   Define the Vehicle
+# ----------------------------------------------------------------------
+
+def vehicle_setup():
+
+    # ------------------------------------------------------------------
+    #   Initialize the Vehicle
+    # ------------------------------------------------------------------
+
+    vehicle = SUAVE.Vehicle()
+    vehicle.tag = 'X57_Maxwell_Mod2'
+
+
+    # ------------------------------------------------------------------
+    #   Vehicle-level Properties
+    # ------------------------------------------------------------------
+
+    # mass properties
+    vehicle.mass_properties.max_takeoff   = 2550. * Units.pounds
+    vehicle.mass_properties.takeoff       = 2550. * Units.pounds
+    vehicle.mass_properties.max_zero_fuel = 2550. * Units.pounds 
+    vehicle.envelope.ultimate_load        = 5.7
+    vehicle.envelope.limit_load           = 3.8 
+    vehicle.reference_area                = 14.76
+    vehicle.passengers                    = 4
+    vehicle.systems.control               = "fully powered"
+    vehicle.systems.accessories           = "commuter"    
+    
+    cruise_speed                          = 135.*Units['mph']    
+    altitude                              = 2500. * Units.ft
+    atmo                                  = SUAVE.Analyses.Atmospheric.US_Standard_1976()
+    freestream                            = atmo.compute_values (0.)
+    freestream0                           = atmo.compute_values (altitude)
+    mach_number                           = (cruise_speed/freestream.speed_of_sound)[0][0] 
+    vehicle.design_dynamic_pressure       = ( .5 *freestream0.density*(cruise_speed*cruise_speed))[0][0]
+    vehicle.design_mach_number            =  mach_number
+    
+    # ------------------------------------------------------------------        
+    #   Main Wing
+    # ------------------------------------------------------------------    
+    wing                                  = SUAVE.Components.Wings.Main_Wing()
+    wing.tag                              = 'main_wing' 
+    wing.sweeps.quarter_chord             = 0.0 * Units.deg
+    wing.thickness_to_chord               = 0.12
+    wing.areas.reference                  = 14.76
+    wing.spans.projected                  = 11.4 
+    wing.chords.root                      = 1.46
+    wing.chords.tip                       = 0.92
+    wing.chords.mean_aerodynamic          = 1.19
+    wing.taper                            = wing.chords.root/wing.chords.tip 
+    wing.aspect_ratio                     = wing.spans.projected**2. / wing.areas.reference 
+    wing.twists.root                      = 3.0 * Units.degrees
+    wing.twists.tip                       = 0.0 * Units.degrees 
+    wing.origin                           = [[2.93, 0., 1.01]]
+    wing.aerodynamic_center               = [3., 0., 1.01] 
+    wing.vertical                         = False
+    wing.symmetric                        = True
+    wing.high_lift                        = True 
+    wing.winglet_fraction                 = 0.0  
+    wing.dynamic_pressure_ratio           = 1.0  
+    airfoil                               = SUAVE.Components.Wings.Airfoils.Airfoil()
+    airfoil.coordinate_file               = '../Vehicles/Airfoils/NACA_63_412.txt'
+    
+    cg_x = wing.origin[0][0] + 0.25*wing.chords.mean_aerodynamic
+    cg_z = wing.origin[0][2] - 0.2*wing.chords.mean_aerodynamic
+    vehicle.mass_properties.center_of_gravity = [[cg_x,   0.  ,  cg_z ]]  # SOURCE: Design and aerodynamic analysis of a twin-engine commuter aircraft
+
+    # Wing Segments
+    segment                               = SUAVE.Components.Wings.Segment()
+    segment.tag                           = 'inboard'
+    segment.percent_span_location         = 0.0 
+    segment.twist                         = 3. * Units.degrees   
+    segment.root_chord_percent            = 1. 
+    segment.dihedral_outboard             = 0.  
+    segment.sweeps.quarter_chord          = 0.
+    segment.thickness_to_chord            = 0.12
+    segment.append_airfoil(airfoil)
+    wing.append_segment(segment)
+
+    segment                               = SUAVE.Components.Wings.Segment()
+    segment.tag                           = 'outboard'
+    segment.percent_span_location         = 0.5438
+    segment.twist                         = 2.* Units.degrees 
+    segment.root_chord_percent            = 1. 
+    segment.dihedral_outboard             = 0. 
+    segment.sweeps.quarter_chord          = 0.
+    segment.thickness_to_chord            = 0.12 
+    segment.append_airfoil(airfoil)
+    wing.append_segment(segment)
+    
+    # Wing Segments
+    segment                               = SUAVE.Components.Wings.Segment()
+    segment.tag                           = 'winglet'
+    segment.percent_span_location         = 0.98
+    segment.twist                         = 1.  * Units.degrees 
+    segment.root_chord_percent            = 0.630
+    segment.dihedral_outboard             = 75. * Units.degrees 
+    segment.sweeps.quarter_chord          = 82. * Units.degrees 
+    segment.thickness_to_chord            = 0.12 
+    segment.append_airfoil(airfoil)
+    wing.append_segment(segment) 
+
+    segment                               = SUAVE.Components.Wings.Segment()
+    segment.tag                           = 'tip'
+    segment.percent_span_location         = 1.
+    segment.twist                         = 0. * Units.degrees 
+    segment.root_chord_percent            = 0.12
+    segment.dihedral_outboard             = 0.
+    segment.sweeps.quarter_chord          = 0.
+    segment.thickness_to_chord            = 0.12
+    segment.append_airfoil(airfoil)
+    wing.append_segment(segment)    
+    
+    # add to vehicle
+    vehicle.append_component(wing)
+
+
+    # ------------------------------------------------------------------        
+    #  Horizontal Stabilizer
+    # ------------------------------------------------------------------       
+    wing                                  = SUAVE.Components.Wings.Wing()
+    wing.tag                              = 'horizontal_stabilizer' 
+    wing.sweeps.quarter_chord             = 0.0 * Units.deg
+    wing.thickness_to_chord               = 0.12
+    wing.areas.reference                  = 2.540 
+    wing.spans.projected                  = 3.3  * Units.meter 
+    wing.sweeps.quarter_chord             = 0 * Units.deg 
+    wing.chords.root                      = 0.769 * Units.meter 
+    wing.chords.tip                       = 0.769 * Units.meter 
+    wing.chords.mean_aerodynamic          = 0.769 * Units.meter  
+    wing.taper                            = 1. 
+    wing.aspect_ratio                     = wing.spans.projected**2. / wing.areas.reference 
+    wing.twists.root                      = 0.0 * Units.degrees
+    wing.twists.tip                       = 0.0 * Units.degrees 
+    wing.origin                           = [[7.7, 0., 0.25]]
+    wing.aerodynamic_center               = [7.8, 0., 0.25] 
+    wing.vertical                         = False
+    wing.winglet_fraction                 = 0.0  
+    wing.symmetric                        = True
+    wing.high_lift                        = False 
+    wing.dynamic_pressure_ratio           = 0.9
+
+    # add to vehicle
+    vehicle.append_component(wing)
+
+
+    # ------------------------------------------------------------------
+    #   Vertical Stabilizer
+    # ------------------------------------------------------------------ 
+    wing                                  = SUAVE.Components.Wings.Wing()
+    wing.tag                              = 'vertical_stabilizer'     
+    wing.sweeps.quarter_chord             = 25. * Units.deg
+    wing.thickness_to_chord               = 0.12
+    wing.areas.reference                  = 2.258 * Units['meters**2']  
+    wing.spans.projected                  = 1.854   * Units.meter  
+    wing.chords.root                      = 1.6764 * Units.meter 
+    wing.chords.tip                       = 0.6858 * Units.meter 
+    wing.chords.mean_aerodynamic          = 1.21   * Units.meter 
+    wing.taper                            = wing.chords.tip/wing.chords.root 
+    wing.aspect_ratio                     = wing.spans.projected**2. / wing.areas.reference 
+    wing.twists.root                      = 0.0 * Units.degrees
+    wing.twists.tip                       = 0.0 * Units.degrees 
+    wing.origin                           = [[6.75 ,0,  0.623]]
+    wing.aerodynamic_center               = [0.508 ,0,0]  
+    wing.vertical                         = True 
+    wing.symmetric                        = False
+    wing.t_tail                           = False
+    wing.winglet_fraction                 = 0.0  
+    wing.dynamic_pressure_ratio           = 1.0
+
+    # add to vehicle
+    vehicle.append_component(wing)
+
+
+    # ------------------------------------------------------------------
+    #  Fuselage
+    # ------------------------------------------------------------------
+    fuselage = SUAVE.Components.Fuselages.Fuselage()
+    fuselage.tag                                = 'fuselage'
+    fuselage.seats_abreast                      = 2.
+    fuselage.fineness.nose                      = 1.6
+    fuselage.fineness.tail                      = 2.
+    fuselage.lengths.nose                       = 60.  * Units.inches
+    fuselage.lengths.tail                       = 161. * Units.inches
+    fuselage.lengths.cabin                      = 105. * Units.inches
+    fuselage.lengths.total                      = 332.2* Units.inches
+    fuselage.lengths.fore_space                 = 0.
+    fuselage.lengths.aft_space                  = 0.
+    fuselage.width                              = 42. * Units.inches
+    fuselage.heights.maximum                    = 62. * Units.inches
+    fuselage.heights.at_quarter_length          = 62. * Units.inches
+    fuselage.heights.at_three_quarters_length   = 62. * Units.inches
+    fuselage.heights.at_wing_root_quarter_chord = 23. * Units.inches
+    fuselage.areas.side_projected               = 8000.  * Units.inches**2.
+    fuselage.areas.wetted                       = 30000. * Units.inches**2.
+    fuselage.areas.front_projected              = 42.* 62. * Units.inches**2.
+    fuselage.effective_diameter                 = 50. * Units.inches 
+
+    # Segment
+    segment                                     = SUAVE.Components.Fuselages.Segment()
+    segment.tag                                 = 'segment_0'
+    segment.percent_x_location                  = 0
+    segment.percent_z_location                  = 0
+    segment.height                              = 0.01
+    segment.width                               = 0.01
+    fuselage.Segments.append(segment)
+
+    # Segment
+    segment                                     = SUAVE.Components.Fuselages.Segment()
+    segment.tag                                 = 'segment_1'
+    segment.percent_x_location                  = 0.007279116466
+    segment.percent_z_location                  = 0.002502014453
+    segment.height                              = 0.1669064748
+    segment.width                               = 0.2780205877
+    fuselage.Segments.append(segment)
+
+    # Segment
+    segment                                     = SUAVE.Components.Fuselages.Segment()
+    segment.tag                                 = 'segment_2'
+    segment.percent_x_location                  = 0.01941097724
+    segment.percent_z_location                  = 0.001216095397
+    segment.height                              = 0.3129496403
+    segment.width                               = 0.4365777215
+    fuselage.Segments.append(segment)
+
+    # Segment
+    segment                                     = SUAVE.Components.Fuselages.Segment()
+    segment.tag                                 = 'segment_3'
+    segment.percent_x_location                  = 0.06308567604
+    segment.percent_z_location                  = 0.007395489231
+    segment.height                              = 0.5841726619
+    segment.width                               = 0.6735119903
+    fuselage.Segments.append(segment)
+
+    # Segment
+    segment                                     = SUAVE.Components.Fuselages.Segment()
+    segment.tag                                 = 'segment_4'
+    segment.percent_x_location                  = 0.1653761217
+    segment.percent_z_location                  = 0.02891281352
+    segment.height                              = 1.064028777
+    segment.width                               = 1.067200529
+    fuselage.Segments.append(segment)
+
+    # Segment
+    segment                                     = SUAVE.Components.Fuselages.Segment()
+    segment.tag                                 = 'segment_5'
+    segment.percent_x_location                  = 0.2426372155
+    segment.percent_z_location                  = 0.04214148761
+    segment.height                              = 1.293766653
+    segment.width                               = 1.183058255
+    fuselage.Segments.append(segment)
+
+    # Segment
+    segment                                     = SUAVE.Components.Fuselages.Segment()
+    segment.tag                                 = 'segment_6'
+    segment.percent_x_location                  = 0.2960174029
+    segment.percent_z_location                  = 0.04705241831
+    segment.height                              = 1.377026712
+    segment.width                               = 1.181540054
+    fuselage.Segments.append(segment)
+
+    # Segment
+    segment                                     = SUAVE.Components.Fuselages.Segment()
+    segment.tag                                 = 'segment_7'
+    segment.percent_x_location                  = 0.3809404284
+    segment.percent_z_location                  = 0.05313580461
+    segment.height                              = 1.439568345
+    segment.width                               = 1.178218989
+    fuselage.Segments.append(segment)
+
+    # Segment
+    segment                                     = SUAVE.Components.Fuselages.Segment()
+    segment.tag                                 = 'segment_8'
+    segment.percent_x_location                  = 0.5046854083
+    segment.percent_z_location                  = 0.04655492473
+    segment.height                              = 1.29352518
+    segment.width                               = 1.054390707
+    fuselage.Segments.append(segment)
+
+    # Segment
+    segment                                     = SUAVE.Components.Fuselages.Segment()
+    segment.tag                                 = 'segment_9'
+    segment.percent_x_location                  = 0.6454149933
+    segment.percent_z_location                  = 0.03741966266
+    segment.height                              = 0.8971223022
+    segment.width                               = 0.8501926505
+    fuselage.Segments.append(segment)
+
+    # Segment
+    segment                                     = SUAVE.Components.Fuselages.Segment()
+    segment.tag                                 = 'segment_10'
+    segment.percent_x_location                  = 0.985107095
+    segment.percent_z_location                  = 0.04540283436
+    segment.height                              = 0.2920863309
+    segment.width                               = 0.2012565415
+    fuselage.Segments.append(segment)
+
+    # Segment
+    segment                                     = SUAVE.Components.Fuselages.Segment()
+    segment.tag                                 = 'segment_11'
+    segment.percent_x_location                  = 1
+    segment.percent_z_location                  = 0.04787575562
+    segment.height                              = 0.1251798561
+    segment.width                               = 0.1206021048
+    fuselage.Segments.append(segment)
+
+    # add to vehicle
+    vehicle.append_component(fuselage)
+
+    #---------------------------------------------------------------------------------------------
+    # DEFINE PROPELLER
+    #---------------------------------------------------------------------------------------------
+    # build network
+    net = Battery_Propeller()
+    net.number_of_propeller_engines  = 2.
+    net.nacelle_diameter             = 42 * Units.inches
+    net.engine_length                = 0.01 * Units.inches
+    net.identical_propellers         = True
+    net.areas                        = Data()
+    net.areas.wetted                 = 0.01*(2*np.pi*0.01/2)
+
+    # Component 1 the ESC
+    esc = SUAVE.Components.Energy.Distributors.Electronic_Speed_Controller()
+    esc.efficiency = 0.95 # Gundlach for brushless motors
+    net.esc        = esc
+
+    # Component 2 the Propeller 
+    prop = SUAVE.Components.Energy.Converters.Propeller()
+
+    prop.number_of_blades       = 2.0
+    prop.freestream_velocity    = 135.*Units['mph']
+    prop.angular_velocity       = 1300.  * Units.rpm
+    prop.tip_radius             = 76./2. * Units.inches
+    prop.hub_radius             = 8.     * Units.inches
+    prop.design_Cl              = 0.8
+    prop.design_altitude        = 12000. * Units.feet
+    prop.design_altitude        = 12000. * Units.feet
+    prop.design_thrust          = 1200.
+    prop.origin                 = [[2.,2.5,0.784]]
+    prop.rotation               = -1
+    prop.symmetry               = True
+    prop.variable_pitch         = True
+
+    prop.airfoil_geometry       =  ['../Vehicles/Airfoils/NACA_4412.txt']
+    prop.airfoil_polars         = [['../Vehicles/Airfoils/Polars/NACA_4412_polar_Re_50000.txt' ,
+                                    '../Vehicles/Airfoils/Polars/NACA_4412_polar_Re_100000.txt' ,
+                                    '../Vehicles/Airfoils/Polars/NACA_4412_polar_Re_200000.txt' ,
+                                    '../Vehicles/Airfoils/Polars/NACA_4412_polar_Re_500000.txt' ,
+                                    '../Vehicles/Airfoils/Polars/NACA_4412_polar_Re_1000000.txt' ]]
+
+    prop.airfoil_polar_stations = [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0]
+    prop                        = propeller_design(prop)
+
+    prop_left = deepcopy(prop)
+    prop_left.origin   = [[2.,-2.5,0.784]]
+    prop_left.rotation = 1
+    
+    net.propellers.append(prop)
+    net.propellers.append(prop_left)
+
+
+    # Component 3 the Battery
+    bat = SUAVE.Components.Energy.Storages.Batteries.Constant_Mass.Lithium_Ion()
+    bat.mass_properties.mass = 500. * Units.kg
+    bat.specific_energy      = 350. * Units.Wh/Units.kg
+    bat.resistance           = 0.006
+    bat.max_voltage          = 500.
+
+    initialize_from_mass(bat,bat.mass_properties.mass)
+    net.battery              = bat
+    net.voltage              = bat.max_voltage
+
+    # Component 4 Miscellaneous Systems
+    sys = SUAVE.Components.Systems.System()
+    sys.mass_properties.mass = 5 # kg
+ 
+    # Component 5 the Motor  
+    motor                         = SUAVE.Components.Energy.Converters.Motor()
+    motor.efficiency              = 0.95
+    motor.gearbox_efficiency      = 1.
+    motor.origin                  = [[2.,  2.5, 0.784]]
+    motor.nominal_voltage         = bat.max_voltage *3/4
+    motor.propeller_radius        = prop.tip_radius
+    motor.no_load_current         = 4.0
+    motor                         = size_optimal_motor(motor,prop)
+    motor.mass_properties.mass    = 10. * Units.kg
+    net.motor                     = motor
+    
+    # append right motor
+    net.propeller_motors.append(motor)
+    
+    # append left motor 
+    motor_left = deepcopy(motor)
+    motor_left.origin = [[2., -2.5, 0.784]]
+    motor_left.rotation = 1
+    net.propeller_motors.append(motor_left) 
+
+    # Component 6 the Payload
+    payload = SUAVE.Components.Energy.Peripherals.Payload()
+    payload.power_draw           = 10. # Watts
+    payload.mass_properties.mass = 1.0 * Units.kg
+    net.payload                  = payload
+
+    # Component 7 the Avionics
+    avionics = SUAVE.Components.Energy.Peripherals.Avionics()
+    avionics.power_draw = 20. # Watts
+    net.avionics        = avionics
+
+    # add the solar network to the vehicle
+    vehicle.append_component(net)
+
+    # ------------------------------------------------------------------
+    #   Vehicle Definition Complete
+    # ------------------------------------------------------------------
+    
+    return vehicle
+
+# ---------------------------------------------------------------------
+#   Define the Configurations
+# ---------------------------------------------------------------------
+
+def configs_setup(vehicle):
+
+    # ------------------------------------------------------------------
+    #   Initialize Configurations
+    # ------------------------------------------------------------------
+
+    configs = SUAVE.Components.Configs.Config.Container()
+
+    base_config = SUAVE.Components.Configs.Config(vehicle)
+    base_config.tag = 'base'
+    base_config.networks.battery_propeller.pitch_command = 0
+    configs.append(base_config)
+
+
+    # done!
+    return configs
